### 2023-12-05 #619 @flor14
quadkeyr: Tools for converting QuadKeys (Microsoft's Bing Maps Tile System) into raster images
Submitting Author Name: Florencia D'Andrea
Submitting Author Github Handle: @flor14
Repository:  https://github.com/Fernandez-Lab-WSU/quadkeyr
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: quadkeyr
Title: Tools for converting QuadKeys used in Microsoft's Bing Maps Tile System into raster images
Version: 0.0.0.9000
    person(given = "Florencia",
           family = "D'Andrea",
           email = "florencia.dandrea@gmail.com",
           comment = c(ORCID = "0000-0002-0041-097X"))
Description: quadkeyr enables the generation of raster images based on QuadKeys, facilitating efficient integration of Bing Maps data into R workflows. In particular, quadkeyr provides support to process and analyze Facebook mobility datasets within the R environment.
    shinytest2,
    bslib (>= 0.5.1),
    DT (>= 0.27),
    ggplot2 (>= 3.4.3),
    leaflet (>= 2.2.0),
    lubridate (>= 1.9.2),
    purrr (>= 1.0.1),
    readr (>= 2.1.4),
    rnaturalearth (>= 0.3.2),
    sf (>= 1.0.14),
    shiny (>= 1.7.4),
    stars (>= 0.6.2),
    tidyr (>= 1.3.0),
    viridis (>= 0.6.4)
```

🗺️  The package offers tools to analyze data reported by QuadKey and convert it to raster images. QuadKeys are strings that encode information about map coordinates and the level of detail of the [Bing Maps Tile System](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).

Anyone trying to analyze data reported in this format. Facebook mobility data, for example, can be reported by QuadKey. This package could help [this StackOverflow user](https://stackoverflow.com/questions/68658693/join-info-about-quadkeys-to-tile-shapefile) or [this other one.](https://gis.stackexchange.com/questions/446516/raster-file-with-all-bing-maps-quadkeys)

The closer R package dealing with this type of data is [`slippymath`](https://github.com/MilesMcBain/slippymath) which has a more general objective. `quadkeyr` is only based on Microsoft Bing Maps Tile System documentation and it is focused on raster creation.
You can read a list of packages with similar functions in the [README references section](https://github.com/Fernandez-Lab-WSU/quadkeyr#references).

✅
✖️
👍🏼  I have fixed all the `pkgcheck` comments.

⚠️  Yes, and I have to mention:
        1. The package coverage is less than 75% but I think that the main reason for that is that the Shiny app included doesn't have tests yet. The backend of that app is all the functions that are tested.
        2. There are also missing 2 notes that still appear after running `check`. The functions `polygon_to_raster` and `format_data` don't have valid roxygen2 examples, and `polygon_to_raster` don't have tests. At least a warning is missing in `create_raster` and some functions could be potentially improved in efficiency and documentation (I would like to improve the technical terminology). Despite this, I think it is a good moment to get some feedback as I will be using the package in the next weeks and I will be available to update it.

📔  There are 3 vignettes, one for each major goal of the package. There are links [on the `README` ](https://github.com/Fernandez-Lab-WSU/quadkeyr) to each of them, not only in the `pkgdown` site. They grow in complexity from goals 1 to 3. I think reading them in order should give a good idea about the package's core functionalities and logic.

❓ Probably yes? After all the packages that were deleted from CRAN, I would appreciate comments about this. I was careful to use as dependencies only `stars` and `sf`, which are packages that will continue to be active.
