### 2023-12-05 a089d7a
Fix vignette rendering issue
