+++ vignettes/get_grid_from_quadkeys.Rmd
@@ -7,7 +7,16 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  warning = FALSE, 
+  message = FALSE, 
+  fig.width=7, 
+  fig.height=7
+)
+```
 
 
 
@@ -17,7 +26,14 @@ vignette: >
 
 #### 1.1 Define the area
 
-
+```{r setup, echo = FALSE}
+library(quadkeyr)
+library(sf)
+library(rnaturalearth)
+library(ggplot2)
+library(dplyr)
+library(stars)
+```
 
 First, we should define the `xmin`, `xmax`, `ymin` and `ymax` that will delimit 
 the area for the QuadKey grid creation.
@@ -25,7 +41,25 @@ the area for the QuadKey grid creation.
 For this example, we have selected `xmin` = -59, `xmax` = -57, `ymin` = -35,
 `ymax` = -34. Let's plot them as points.
 
-![plot of chunk unnamed-chunk-2](figure/unnamed-chunk-2-1.png)
+```{r echo=FALSE}
+
+
+ggplot() + 
+  geom_sf(data = ne_countries(returnclass = 'sf'),
+          fill = 'beige') +
+  geom_point(aes(x = -59, y = -35 ), color = 'purple') +
+  geom_point(aes(x = -57, y = -34), color = 'purple') +
+  coord_sf(xlim = c(-60, -56),
+            ylim = c(-36, -33), expand = FALSE) +
+    theme_minimal()  +
+  theme(panel.background = element_rect(fill = "lightblue"),
+        panel.ontop = FALSE,
+        panel.grid.major = element_blank(), 
+        panel.grid.minor = element_blank()) +
+  ylab("Latitude") + 
+  xlab("Longitude")
+
+```
 
 #### 1.2 Select the level of detail
 
@@ -41,8 +75,7 @@ each element of the grid
   
 3. `grid$num_cols` the number of columns of the grid.
 
-
-```r
+```{r cqg}
 grid <-  create_qk_grid(
                xmin = -59,
                xmax = -57,
@@ -50,13 +83,7 @@ grid <-  create_qk_grid(
                ymax = -34,
                level = 12)
 head(grid$data)
-#>   tileX tileY      quadkey
-#> 1  1376  2473 210321302002
-#> 2  1376  2472 210321302000
-#> 3  1376  2471 210321300222
-#> 4  1376  2470 210321300220
-#> 5  1376  2469 210321300202
-#> 6  1376  2468 210321300200
+
 ```
 
 #### 1.3 Get the grid coordinates from the QuadKeys
@@ -64,26 +91,34 @@ head(grid$data)
 The coordinates are extracted from the QuadKeys 
 using the function `extract_qk_coords()`
 
-
-```r
+```{r grid}
 grid_coords <- extract_qk_coord(data = grid$data)
 head(grid_coords)
-#> Simple feature collection with 6 features and 3 fields
-#> Geometry type: POINT
-#> Dimension:     XY
-#> Bounding box:  xmin: -59.0625 ymin: -34.958 xmax: -59.0625 ymax: -34.59704
-#> Geodetic CRS:  WGS 84
-#>   tileX tileY      quadkey                   geometry
-#> 1  1376  2473 210321302002   POINT (-59.0625 -34.958)
-#> 2  1376  2472 210321302000 POINT (-59.0625 -34.88593)
-#> 3  1376  2471 210321300222  POINT (-59.0625 -34.8138)
-#> 4  1376  2470 210321300220 POINT (-59.0625 -34.74161)
-#> 5  1376  2469 210321300202 POINT (-59.0625 -34.66936)
[…]
