### 2023-12-05 65762e2
Fix problem map rendering - vignette
