+++ docs/packages.bib
@@ -1,90 +1,0 @@
-@Manual{R-devtools,
-  title = {devtools: Tools to Make Developing R Packages Easier},
-  author = {Hadley Wickham and Jim Hester and Winston Chang and Jennifer Bryan},
-  year = {2022},
-  note = {R package version 2.4.5},
-  url = {https://CRAN.R-project.org/package=devtools},
-}
-
-@Manual{R-dplyr,
-  title = {dplyr: A Grammar of Data Manipulation},
-  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
-  year = {2023},
-  note = {R package version 1.1.2},
-  url = {https://CRAN.R-project.org/package=dplyr},
-}
-
-@Manual{R-ggplot2,
-  title = {ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics},
-  author = {Hadley Wickham and Winston Chang and Lionel Henry and Thomas Lin Pedersen and Kohske Takahashi and Claus Wilke and Kara Woo and Hiroaki Yutani and Dewey Dunnington},
-  year = {2023},
-  note = {R package version 3.4.3},
-  url = {https://CRAN.R-project.org/package=ggplot2},
-}
-
-@Manual{R-rnaturalearth,
-  title = {rnaturalearth: World Map Data from Natural Earth},
-  author = {Philippe Massicotte and Andy South},
-  year = {2023},
-  note = {R package version 0.3.2},
-  url = {https://docs.ropensci.org/rnaturalearth/
-https://github.com/ropensci/rnaturalearth},
-}
-
-@Manual{R-sf,
-  title = {sf: Simple Features for R},
-  author = {Edzer Pebesma},
-  year = {2023},
-  note = {R package version 1.0-14},
-  url = {https://CRAN.R-project.org/package=sf},
-}
-
-@Manual{R-stars,
-  title = {stars: Spatiotemporal Arrays, Raster and Vector Data Cubes},
-  author = {Edzer Pebesma},
-  year = {2023},
-  note = {R package version 0.6-2},
-  url = {https://CRAN.R-project.org/package=stars},
-}
-
-@Book{ggplot22016,
-  author = {Hadley Wickham},
-  title = {ggplot2: Elegant Graphics for Data Analysis},
-  publisher = {Springer-Verlag New York},
-  year = {2016},
-  isbn = {978-3-319-24277-4},
-  url = {https://ggplot2.tidyverse.org},
-}
-
-@Book{sf2023,
-  author = {Edzer Pebesma and Roger Bivand},
-  title = {{Spatial Data Science: With applications in R}},
-  year = {2023},
-  publisher = {{Chapman and Hall/CRC}},
-  url = {https://r-spatial.org/book/},
-  doi = {10.1201/9780429459016},
-}
-
-@Article{sf2018,
-  author = {Edzer Pebesma},
-  title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
-  year = {2018},
-  journal = {{The R Journal}},
-  doi = {10.32614/RJ-2018-009},
-  url = {https://doi.org/10.32614/RJ-2018-009},
-  pages = {439--446},
-  volume = {10},
-  number = {1},
-}
-
-@Book{stars2023,
-  title = {{Spatial Data Science: With applications in R}},
-  author = {Edzer Pebesma and Roger Bivand},
-  year = {2023},
-  publisher = {Chapman and Hall/CRC},
-  address = {London},
-  url = {https://r-spatial.org/book/},
-  pages = {352},
-  doi = {10.1201/9780429459016},
-}
-
