+++ vignettes/get_grid_from_quadkeys.Rmd
@@ -41,7 +41,7 @@ the area for the QuadKey grid creation.
 For this example, we have selected `xmin` = -59, `xmax` = -57, `ymin` = -35,
 `ymax` = -34. Let's plot them as points.
 
-```{r echo=FALSE}
+```{r pointsplot, echo=FALSE}
 
 
 ggplot() + 
@@ -97,7 +97,7 @@ head(grid_coords)
 ```
 We can visualize the points in the map to understand better the results.
 
-```{r echo = FALSE}
+```{r mapgrid, echo = FALSE}
 
 
 min =st_sfc(st_point(c(-59,-35)), crs = 4326)
@@ -132,7 +132,7 @@ polygrid = grid_to_polygon(grid_coords)
 head(polygrid)
 
 ```
-```{r echo = FALSE}
+```{r map, echo = FALSE}
 
 min =st_sfc(st_point(c(-59,-35)), crs = 4326)
 max = st_sfc(st_point(c(-57,-34)), crs = 4326)
@@ -177,7 +177,7 @@ Let's generate the raster. The `data_provided` dataset consists of QuadKeys
 linked to a variable value. While this dataset contains QuadKeys within our
 defined study area and detail level, not all grid QuadKeys are present.
 
-```{r}
+```{r datap}
 data('data_provided')
 head(data_provided)
 
@@ -199,7 +199,7 @@ head(data_raster)
 ```
 Now, we can use the dataframe to create the raster.
 
-```{r}
+```{r raster}
 raster <-  create_raster(template = data_raster,
               nx = grid$num_cols +1,
               ny = grid$num_rows +1,
@@ -215,7 +215,7 @@ raster
 ```
 
 
-```{r echo = FALSE}
+```{r plot, echo = FALSE}
 
 
 min =st_sfc(st_point(c(-59,-35)), crs = 4326)
