### 2023-12-06 c5adb89
Add arguments to read_all_files
