### 2023-12-06 c9aa39a
Avoid reading empty files in read_all_files(). Update vignette
