### 2023-12-06 4fe0920
Add col_names argument in read_all_files. Update vignette.
