+++ vignettes/create_rasters_from_grid.Rmd
@@ -44,9 +44,9 @@ change.
 ```{r eval=FALSE, include=TRUE}
 
 files <- read_all_files(path_to_csvs = "../geocovid/data/rasters/city/",
-                        colnames = c("lat", "lon", "quadkey", "date_time", 
-                                     "n_crisis", "percent_change"),
-                        coltypes = "dddTcc")
+                        colnames = c("lat", "lon", 
+                                     "quadkey", "date_time", 
+                                     "n_crisis", "percent_change"))
 
 files
 
