+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   create_rasters_from_grid: create_rasters_from_grid.html
   get_grid_from_quadkeys: get_grid_from_quadkeys.html
   quadkey_conversion: quadkey_conversion.html
-last_built: 2023-12-06T21:04Z
+last_built: 2023-12-06T21:22Z
 
