### 2023-12-06 2a2a097
Change col_type argument read_all_files and update vignette
