### 2023-12-06 f112c2a
Fix typo read_all_files()
