### 2023-12-07 b5d0d76
Add shiny tests and move app to inst/
