+++ man/qkmap_app.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/qkmap_app.R
+\name{qkmap_app}
+\alias{qkmap_app}
+\title{Launches the Shiny App}
+\usage{
+qkmap_app()
+}
+\description{
+This function launches the Shiny app.
+}
