+++ NAMESPACE
@@ -18,6 +18,7 @@ export(missing_combinations)
 export(pixelXY_to_latlong)
 export(pixelXY_to_tileXY)
 export(polygon_to_raster)
+export(qkmap_app)
 export(quadkey_to_latlong)
 export(quadkey_to_tileXY)
 export(read_all_files)
