### 2023-12-08 #619 @jhollist
@flor14 Thank you for the submission and apologies for the few days that went by prior to my response.  Had a high priority thing at work that I had to focus on!

I will take a closer look at your submission early next week.  One thing we will need to get figured out is the R CMD Check failures that our build system is seeing.  I noticed that the builds on your repo seem to be passing just fine.  Stay tuned and thanks again for the submission.
