### 2023-12-08 #619 @flor14
Thank you @jhollist, don't worry. I have increased the code coverage to 84% adding tests for the Shiny app, so that is solved. About the checks, I am unsure what could be the issue, as all the tests are passing for me as you can see in the repo, but I will try to figure it out.
