### 2023-12-11 #619 @flor14
@jhollist the RMD checks are failing because there was missing this image, which now is included in the repo: https://github.com/Fernandez-Lab-WSU/quadkeyr/blob/main/figure/unnamed-chunk-2-1.png
This is the error that the bot reported:
```
    creating vignettes ... ERROR
    --- re-building ‘create_rasters_from_grid.Rmd’ using rmarkdown
    --- finished re-building ‘create_rasters_from_grid.Rmd’

--- re-building ‘get_grid_from_quadkeys.Rmd’ using rmarkdown
File figure/unnamed-chunk-2-1.png not found in resource path
```

I hope that if you run the RMD Checks again, they should be ✅ without issues.
Also, I will continue using the package these days and probably improve things, let me know if you want me to do a first release for the review.
