### 2023-12-11 #619 @jhollist
@flor14 Thanks for tracking down that issue!  That was really helpful.

Couple of thoughts for you.

1.) I re-ran the checks and the only thing that comes up now is related to the new function you added, `qkmap_app`.  I am guessing you are still in progress on that.
2.) If you are going to be working on the package in the next few days I would prefer to see you reach a somewhat stable release that we can use as the basis for the review.
3.) If you want to run the checks on your own, you can use the action at https://github.com/ropensci-review-tools/pkgcheck-action.

In the meantime I will look closer at the package and vignettes.
