### 2023-12-12 cc8526a
Fix path to app in qkmap_app()
