### 2023-12-12 c3fcf73
Fix minor details. Make example qkmap_app don't run to pass checks
