+++ R/qkmap_app.R
@@ -4,15 +4,15 @@
 #' 
 #' @export
 #' 
-#' @example 
-#' 
-#' qkmap_app()
-#' 
+#' @examples 
+#' \dontrun{
+#' qkmap_app() 
+#' } 
 qkmap_app <- function() {
   # Get the path to the Shiny app directory
   shiny_app_path <- system.file("app",
                                 package = "quadkeyr")
   
   # Launch the Shiny app
-  shiny::runApp(appDir = "inst/app/")
+  shiny::runApp(shiny_app_path)
 }
