### 2023-12-12 10376c7
Try to fix bug in examples - qkmap_app()
