### 2023-12-12 #619 @flor14
@jhollist `pkgcheck` is passing now, I have added the GitHub action to the repository.
