### 2023-12-13 #619 @flor14
Yes, I think it is ready. The issue was that my changes were passing the `R CMD checks`, so I didn't anticipate issues before the review started. Now I learned that `pkgcheck` works differently and seems to be more strict.
If there is any new change (even if it is small)  I will retain it to make all of them together after the review, to avoid any potential issue.
Thank you @jhollist
