### 2023-12-19 #619 @emilyriederer
Hi @flor14 ! It's great to meet you. Thanks for your submission.

As @jhollist mentioned, I may not be able to engage more fully until next week, but I thought I'd send over a few preliminary observations.

I think this is a compelling project and is close to being ready for reviewers. On a cursory review of the automated outputs, I just had a few small requests to touch up before we hand this off:

- Could you add a sentence to the documentation describing what QuadKeys are? This might help many people with relevant problems realize your package is a solution even if they have never used that tech before (and similarly, it will help our reviewers understand)
- `lintr` noted a number of `library()` calls in your package code. Could you look to remove these, please? All packages in your `Imports` can be assumed to be loaded
- Ideally, I'd love to see the R CMD checks resolved before passing this along. Do you have thoughts on:
    + For (2), could you remove dependencies from the `DESCRIPTION` if they are not used in the core package code? (versus used in things like vignettes)
    + For (3), could you import `.data` from `rlang` as described [here](https://cran.r-project.org/web/packages/dplyr/vignettes/in-packages.html)? That should clear things up

Thank you!
