### 2023-12-20 #619 @flor14
Hello, @emilyriederer, nice to meet you as well.

Including more low-level documentation will be a good idea, I have been paying attention mostly to the function documentation so far.
Also, I realized that 2 notes were pending in the RMD check, but I had some issues when removing the vignette dependencies so I kept them. The same with the `library()` calls that are probably referring to the vignettes. I will try to go more in-depth to fully solve that. I will write a message in this thread to let you know when is done. Thank you!
