### 2023-12-22 #619 @emilyriederer
Hi @flor14 ! Thanks so much for the thoughts. I think with that context we are even closed to moving forward here.

- Once you've updated the super high-level README with a bit more details to help new-to-QuadKeys readers understand the value, please ping me and we will start looking for reviewers
- Given what you shared about the CMD check, I think we can move to reviewers before this is "fixed" and it's also not direly critical unless/until you decide to submit to CRAN. That said, I know the issue you speak so I'll detail one solution below if you're interested

----

[This section](https://r-pkgs.org/vignettes.html#sec-vignettes-eval-option) of the *R Packages* book outlines an approach to solve your problem. In short, you can move such dependencies to Suggests versus Depends in your DESCRIPTION and then consider evaluating conditional on package availability
