### 2023-12-22 f6c0838
Add high-level doc and solves 2check notes based on ropensci/software-review#619
