+++ README.Rmd
@@ -13,8 +13,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# quadkeyr <img src="man/figures/logo.png" align="right" height="250" />
-
+# `quadkeyr` <img src="man/figures/logo.png" align="right" height="200" />
 
 <!-- badges: start -->
 [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental) [![test-coverage](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/test-coverage.yaml) [![R-CMD-check](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/check-standard.yaml/badge.svg)](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/check-standard.yaml) [![Codecov](https://img.shields.io/codecov/c/github/Fernandez-Lab-WSU/quadkeyr)](https://codecov.io/gh/Fernandez-Lab-WSU/quadkeyr)
@@ -22,11 +21,31 @@ knitr::opts_chunk$set(
 
 ⚠️ This package is under review
 
-The `quadkeyr` R package presents a comprehensive toolkit tailored for generating 
-raster images from QuadKeys within [Microsoft's Bing Maps Tile System](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system). 
-Designed to integrate Bing Maps data into R workflows, this package facilitates the creation 
-of QuadKey grids and rasters and introduces specialized functions for
+
+### What are QuadKeys in Tile Maps?
+
+Tile maps divide the Earth's surface into a grid of tiles, with each tile
+corresponding to a specific geographic area at various zoom levels. Different 
+zoom levels in a tile map display varying levels of detail.
+
+QuadKeys represent a location on a map by encoding its hierarchical
+spatial position as a sequence of characters. They provide an efficient method 
+to address and retrieve specific map tiles, facilitating rapid display within 
+mapping applications such as [Microsoft Bing Maps](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
+
+![](quadkeys.jpg)
+  
+  
+*The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage.*
+
+### What can this package do for you?
+The `quadkeyr` R package presents a comprehensive toolkit tailored for
+generating raster images from data reported by QuadKeys within [Microsoft's Bing Maps Tile System](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system). 
+Designed to integrate Bing Maps data into R workflows,
+this package facilitates the creation of QuadKey grids and raster images
+and introduces specialized functions for
 the processing of [Facebook Mobility data](https://dataforgood.facebook.com/).
+
 
 ---
 
@@ -46,7 +65,6 @@ analyze Facebook mobility data.
   
 4. Introduce a QuadKey map application enabling users to validate function 
 outcomes and offering tools for visualizing the generated data.
-
 
 ## Installation
 
