+++ NAMESPACE
@@ -25,3 +25,4 @@ export(read_all_files)
 export(regular_qk_grid)
 export(tileXY_to_pixelXY)
 export(tileXY_to_quadkey)
+importFrom(rlang,.data)

[…]
