### 2023-12-22 f47d234
Fix pending note check - Remove unnecesary files at top level
