+++ README.Rmd
@@ -33,7 +33,10 @@ spatial position as a sequence of characters. They provide an efficient method
 to address and retrieve specific map tiles, facilitating rapid display within 
 mapping applications such as [Microsoft Bing Maps](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
 
-![](quadkeys.jpg)
+```{r echo=FALSE, fig.align='center'}
+knitr::include_graphics("man/figures/quadkeys.jpg")
+```
+
   
   
 *The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage.*
