### 2023-12-22 #619 @flor14
Thank you @emilyriederer

- The documentation has been improved in the README (I should have said high-level doc)
- `rlang` was added as a dependency
- Some packages were moved to `Suggest`. Some of them are used in an internal Shiny app as well, not only vignettes.
👉🏼  Fernandez-lab-WSU/quadkeyr@f6c0838
Also, I solved the last note removing some files at the top-level directory and rebuilding the `pkgdown` website.
👉🏼  Fernandez-lab-WSU/quadkeyr@f47d234

Now everything should be passing without notes.

In case a bit of context helps as well, this package is part of a "trilogy". I am developing a [Shiny app](https://piliffq.shinyapps.io/geocovid_beta/) that reads and visualizes the raster images generated with `quadkeyr` and a [website](https://fernandez-lab-wsu.github.io/geocovid_bsas/) with all the data analysis necessary for that app. I am actively working on making the three of them coherent now, the idea is to keep the reusable functions in `quadkeyr` to ensure reproducibility.

I hope you have happy holidays
