### 2023-12-22 #619 @emilyriederer
### Editor checks:

- [x ] **Documentation**: The package has sufficient documentation available online (README, pkgdown docs) to allow for an assessment of functionality and scope without installing the package. In particular,
    - [x ] Is the case for the package well made?
    - [x ] Is the reference index page clear (grouped by topic if necessary)? **There could be more grouping here; however, the clarity of function naming makes the page sufficiently easy to navigate**
    - [x ] Are vignettes readable, sufficiently detailed and not just perfunctory? **Yes, vignettes are complete and well-narrated. The titles could be improved so they are more than function names to improve navigability.**
- [ x] **Fit**: The package meets criteria for [fit](https://devguide.ropensci.org/policies.html#package-categories) and [overlap](https://devguide.ropensci.org/policies.html#overlap).
- [x ] **Installation instructions:** Are installation instructions clear enough for human users?
- [x ] **Tests**: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)?
- [x ] **Contributing information**: Is the documentation for contribution clear enough e.g. tokens for tests, playgrounds?
- [x ] **License:** The package has a CRAN or OSI accepted license.
- [x ] **Project management**: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled?
---

#### Editor comments

This package seems very well-motivated and well-polished for reviewers. The other nice-to-haves on pkgdown might be:

- Add more organization to the References page by tweaking the YAML template [per the example here](https://pkgdown.r-lib.org/articles/pkgdown.html#reference)
- Updating vignette names to be something more readable than function names

However, these are both very minor points and neither is an impediment to moving forward with the review

---
