+++ NEWS.md
@@ -1,1 +1,3 @@
+# quadkeyr 0.0.0.9000
 
+* Added a `NEWS.md` file to track changes to the package.
