### 2023-12-26 #619 @emilyriederer
Hi @flor14 - I'm delighted to introduct @mpaulacaldas as our first reviewer. Maria has substantial experience in geospatial analysis and R package development, including with quadkey datasets, so she will bring a fantastic perspective. I'm still seeking a second reviewer
