### 2023-12-26 #619 @emilyriederer
Thank you, @mpaulacaldas ! That said, our hardworking bot doesn't know that it's the holidays, and I know for many this isn't a typical workweek. Given that (and that I still need one more reviewer!) please feel free to take an extra week if needed 🙂
