### 2023-12-27 #619 @flor14
Thank you @mpaulacaldas for being a reviewer. It is great that you have used quadkeys before, I hope the package is useful and of interest!
