### 2023-12-30 #619 @emilyriederer
Hi @flor14 - I'm please to introduce our second reviewer @vincentvanhees. Dr. Van Hees is an experienced R developer and consultant who was previously written and maintained geospatial R packages. I think this experience will bring great insights to the `quadkeyr` review
