### 2024-01-03 #619 @vincentvanhees
Here is my review :) Let me know if anything is unclear.

- **Briefly describe any working relationship you have (had) with the package authors.** => None

---

I enjoyed reviewing this package, it is well designed. Even though raster data and quadkeys are not my expertise, I was able to figure out what the package is doing.

**1. Vignettes and README:**
  - The vignette title fields and filenames are not intuitive to me. For example, `create_rasters_from_grid` is about creating rasters, but `get_grid_from_quadkeys` also shows how to create a raster. If I understand correctly, the high level appeal of the package is to facilitate working with facebook quadkey mobility data or to create your own quadkeys and matching grid and raster, but that is not reflected in the titles. Can you come up with more catchy titles? Note that if this would be a CRAN package then the README would not be visible on the CRAN page and users typically first go to one of the vignettes. Therefore, it may be good to rethink the vignette names and titles a bit. How about:

current filename | new title | new filename
----------|----------|---------------------
create_rasters_from_grid.Rmd | From facebook quadkey data to raster |  from_facebook_quadkey_data_to_raster.Rmd
get_grid_from_quadkeys.Rmd | Create your own quadkeys and matching grid and raster | create_your_own_quadkeys.Rmd
quadkey_conversion.Rmd | Bing Maps Tile System Functions | bing_maps_tile_tystem_functions.Rmd

  - Vignette titles should to be entered in the title field at the top and not as a first "#" item further down. At least it does not look good at the moment to have the underscored title at the top of the page:

  - Consider starting each vignette with a link to the README just in case a visitor arrives first at the vignette and is in need for a bit more introduction.
  - Specific comments on get_grid_from_quadkeys.Rmd:
    - Explain what `tileX` and `tileY` mean, I suppose this is an index number relative to upper left corner? If yes, a short sentence to state that would be helpful.
    - 1.3 I would start by explaining that this is about coordinates of the upper-left corner of each square in the grid. In that way the text should be easier to follow as it took me until the end of the paragraph before I realised what coordinates it is refering to.
    - There are two 1.4 paragraphs? Maybe use auto-numbering of paragraphs to avoid issues like this ("number_sections: true")
    - Second 1.4 Consider adding explanation of how data was created? Even if it is fake data you could describe whether its format was created after the format of the facebook mobility data.
  - specific comments on create_rasters_from_grid.Rmd
    - Step 1: Add one sentence to provide introduction to the nature of mobility data. Currently it assumes that the reader is familiar with the format or will visit the external website. Just one sentence to clarify that Facebook mobility data is stored in csv format across a number of subfolders may be enough to make the vignette more accessible.
    -	Step 1: You do not clarify that read_all_files is a quadkeyr function, given that function name is unspecific, I am wondering whether it may help to have a sentence to clarify that read_all_files is part of this package?
    -	Step 1: Object name "files" is confusing as the object itself includes not a vector with filenames but the actual data you will work with further down. So, I would rename this to fb_data or qd_data or something along those lines.
    -	Step 2: Typos "have data al least once".
  - specific comments quadkey_conversion.Rmd:
    -	Needs more introduction at the top to clarify context. For example, why would someone be interested in functions? Maybe the text under item 3 is part of the clarification, so possibly some of that needs to be mentioned at the beginning of the vignette?
    - README: Reconsider clarity of section "What can this package do for you?" after going through my comments on the vignettes above.

**2. R functions:**
  - apply_weekly_lag.R:
    - Parameter 'data' is not documented, because "a dataset" is not specific. Maybe you can copy paste the documentation from other functions where you did make a specific reference to object class and structure.
    - Column names `n_crisis_lag_7` and `n_crisis_percent_7` lack explanation.
    - Indentation missing (Ctrl+A, Ctrl+I would fix this).
  - create_qk_grid.R:
    - Consider putting the values on line 37-40 as default values in the function definition (line 30). In that way you will catch when user forgets to specify a coordinate value.
  - extract_coords.R:
    - Line 36: Shows comment "Check that the data has the correct dimensions..." but the code actually include a check on the number of dimensions, so this comment needs to be omitted?
  - grid_to_polygon.R:
    - Why is function at the bottom commented out? If you would like to keep it there then I would suggest adding a comment in English to clairfy why it is there.
  - latlong_to_qk_functions.R:
    - Usage of functions `latlong_to_pixelXY` and `pixelXY_to_tileXY` on lines 160-168 are duplicated calculations. It seems to me that you can simplify this by:

```
data[i, c('pixelX', 'pixelY')] <- latlong_to_pixelXY(data$lat[i],
                                                       data$lon[i],
                                                       data$level[i])
data[i, c('tileX', 'tileY')]  <-  pixelXY_to_tileXY(data$pixelX[i],
                                                    data$pixelY[i])
```
    -	Given the critical role of these function, you may want to consider including some input checks, e.g. pixel needs to be a positive integer value in a certain range and lon and lat need to be numeric values in certain range. Although, I cannot oversee whether there may be better places to perform such a check more centrally.
    - A comment to clarify where the number 256 comes from would be good. At least for me as outsider it a bit scary to see numbers in code without knowing where they come from and whether or not they constrain the functionality of the package. If 256 is not a universal standard, then consider adding it as a parameter with 256 as default value such that package users can change it. If it is then explain what it is.
  - map_functions.R
    - Similar to my comment above, clarification on why `0.0254` and `256` would be good
  - polygon_to_raster.R
    - Code indentation missing
    - Examples shows "#TBD"?
  - qk_to_latlong_functions.R:
    - Same possible redundancy as latlong_to_qk_functions.R (see above).
    - There is a function commented out at the end. If this is important then add comment to clarify why it needs to be there.
    - Comments in Spanish, not an issue but just a heads up that you alternate between English and Spanish.
  - read_and_format_files.R:
    - read_all_files:
      - Consider renaming to `read_facebook_mobility_files` to make it more specific to its intended functionality.
      - Function does not check whether there are actual files to read. Consider adding a check and error message if no files are found.
       - Line 66: Is it necessary to show this message or can this be skipped to avoid cluttering console output?
    - format_data.R:
      - Line 76: Description is not clear English, please check.
      - Line 101: It looks like time is not actually the time, but only the hour. Why not call it hour to avoid misunderstandings? Is it possible that the user has minute level data? If yes, should they be informed that minutes are ignored?
  - regular_grid.R:
     - Missing indentation

**3. Test function:**
  - `test_read_and_format_files.R`: Only tests the formatting for the files and not the reading. It may be clearer to name the file just `test-format-files.R` to avoid the impression that reading files is covered.

**4. Shiny app:**
  - As new user it is not obvious to me what the possible value range for quadkeys are that I can enter. Maybe good to provide reactive user feedback when value is incorrect or explain in the UI what values are allowed?
  -	It is not clear to me why the app code has to be in the `inst/` folder. Is this a new convention in R package development? What is the concern about putting the app code in the R folder just like all other R code?

**5. General / Other comments:**
  - DESCRIPTION file: R dependency >= 2.10 but `|>` is used in the code which was not in R back then, so presumably this needs to be a more recent R dependency?
  - Package LICENSE is not well recognised by github (see repo package right column). There is a double license file, so maybe that explains it?

  - DESCRIPTION file: Package Title is not fully in Title case and in Description field the package is not between quotes, see https://devguide.ropensci.org/building.html#crangotchas

**6. Suggestions (no need to respond):**
  - I see a lot of for loops. If package is slow on real life datasets, which I cannot verify, then there seems opportunity speeding up the code by replacing those loops by apply calls or similar.
  -  I see a lot of empty lines. I am not able to find clear guidance on this in the [guide](https://devguide.ropensci.org/building.html), but personally, I would omit most of the empty lines to make the code more compact and easier to navigate.
