### 2024-01-08 c34a2a6
Change vignettes name in website (ropensci/software-review#619)
