### 2024-01-11 b3f1fbc
Update text vignettes + improve image rendering (ropensci/software-review#619)
