+++ vignettes/1_bing_map_tile_system_functions.Rmd
@@ -21,11 +21,70 @@ library(quadkeyr)
 > Please, visit the [README](https://fernandez-lab-wsu.github.io/quadkeyr/) 
 > for general information about this package
 
-These functions are the R version of the ones described in [Bing Maps Tile System webpage](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
+Most of the function and arguments names conserved from the documentation.
+`quadkeyr` functions are the R version of the ones described in [Bing Maps Tile System webpage](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
 
-### 1. If you have map coordinates, convert them to QuadKey numbers
+Tile maps are composed of pixels that are grouped into tiles. Later, the tiles
+are converted to QuadKeys to optimize map performance, among other benefits 
+described in detail in the documentation.
 
-#### 1.1 Convert lat/long coordinates to pixel XY coordinates
+Pixels and tiles are expressed as two-dimensional coordinates - (`pixelX`, 
+`pixelY`) and (`tileX`, `tileY`) - but Quadkeys are one-dimensional
+numeric strings. This is important to understand how the conversion works.
+
+Each geographic pair of coordinates (latitude/longitude) will belong to 
+a specific pixel referenced by coordinates (`pixelX`, `pixelY`) 
+for each level of detail. In Fig. 1, you can see pixels (0, 0) and (2047, 2047)
+for level of detail 3. The `quadkeyr` function `latlong_to_pixelXY`, converts 
+geographic coordinates (latitude/longitude) in pixel coordinates for each
+level of detail.
+
+
+```{r pixels, echo = FALSE, out.width= "50%", fig.align='center', fig.cap= "Fig 1. Pixels (0, 0) and (2047, 2047) for a map with level of detail 3. Image extracted from Microsoft's Bing Maps Tile System webpage."}
+knitr::include_graphics('bing_pixel.jpg')
+```
+
+
+A tile is comprised of 256x256 pixels. The tile coordinates are visually
+represented in Fig. 2. These coordinates will ultimately be converted 
+into one-dimensional QuadKeys.
+
+For instance, a pixel for level of detail 3 represented by
+the coordinates `pixelX` = 255 and `pixelY` = 12 is part of the tile with 
+coordinates `tileX` = 0 and `tileY` = 0. 
+The pixel with coordinates (0,0) belongs to tile (0,0), and the 
+pixel (2047, 2047) is part of tile (7,7). You can verify this using 
+the function `pixelXY_to_tileXY` and by comparing Fig. 1 and Fig. 2.
+
+```{r tiles, echo = FALSE, out.width= "50%", fig.align='center', fig.cap= "Fig 2. Tile coordinates. Image extracted from Microsoft's Bing Maps Tile System webpage."}
+knitr::include_graphics('bing_tiles.jpg')
+```
+
+To convert from tile XY coordinates to QuadKeys for a particular
+level of detail, you can use the function `tileXY_to_quadkey`. 
+The digits of the QuadKey number will correspond to the level of detail
+of the map. For example, for the tile (4, 7) of level of detail 3, the
+QuadKey number will be `322`.
+
+This pertains to the conversion from geographic coordinates (latitude/longitude)
+to QuadKeys. However, when converting in the opposite direction, 
+from QuadKeys to geographic coordinates (latitude/longitude), 
+the final pair of latitude/longitude corresponds to 
+the upper-left corner coordinates of the tile and pixel associated with
+your original QuadKey.
+
+To understand this in more detail, have a look to the functions and then read
+section 3.
+
+### 1. If you have geographic map coordinates, convert them to QuadKey numbers
+
+**Geographic coordinates** -> **pixel coordinates** ->
[…]
