+++ vignettes/3_from_facebook_quadkey_data_to_raster.Rmd
@@ -35,7 +35,10 @@ library(stars)
 This section focuses on creating a raster from QuadKey data formatted as 
 provided by [Facebook Mobility data](https://dataforgood.facebook.com/).
 
-### Step 1: read all the `csv` files from a folder and format them
+Facebook mobility data for each location and level of detail is stored in
+a separate `csv` file for each day and time.
+
+### 1. read all the `csv` files from a folder and format them
 
 All this files are for the same area and level of detail, but dates and times
 change.
@@ -75,13 +78,13 @@ head(files)
 ```
 
 
-### Step 2: Convert the Quadkey grid to lat/long coordinates
+### 2. Convert the Quadkey grid to latitude/longitude coordinates
 
-Even if these files correspond to the same area of analysis, they can vary in the
-number of QuadKeys that are reported.
+Even if these files correspond to the same area of analysis, they can vary in 
+the number of QuadKeys that are reported.
 
-That is why we select from all the files all the QuadKeys that have data al least
-once and convert them to coordinates.
+That is why we select from all the files all the QuadKeys that have data at 
+least once and convert them to coordinates.
 
 ```{r}
 
@@ -112,7 +115,7 @@ ggplot() +
   xlab("Longitude")
 ```
 
-### Step 3: Complete the grid 
+### 3. Complete the grid 
 
 Some of the QuadKeys are missing, we can't consider this a regular grid. 
 
@@ -147,7 +150,7 @@ ggplot() +
   xlab("Longitude")
 ```
 
-### Step 4: Create the polygons
+### 4. Create the polygons
 
 
 ```{r}
@@ -188,7 +191,7 @@ polyvar7 <-  polyvar |>
 ```
 
 
-### Step 5: Create the rasters for the variables and data involved.
+### 5. Create the rasters for the variables and data involved.
 
 The rasters are going to be created automatically for each day and time reported.
 Each raster will be created as `<filename>_<date>_<time>.tif`.
