### 2024-01-11 6716486
Update package installation to remotes - README (ropensci/software-review#619)
