### 2024-01-11 9c81a05
Rename functions (ropensci/software-review#619) and change vignette names to pass checks
