+++ man/format_fb_data.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/read_and_format_files.R
-\name{format_data}
-\alias{format_data}
+\name{format_fb_data}
+\alias{format_fb_data}
 \title{Format the data}
 \usage{
-format_data(data)
+format_fb_data(data)
 }
 \arguments{
 \item{data}{A dataframe with a quadkey, date_time, country columns and
@@ -20,9 +20,9 @@ data.
 \examples{
 
 #data(onefile)
-#format_data(data = onefile)
+#format_fb_data(data = onefile)
 
 }
 \seealso{
-\code{\link{read_all_files}}
+\code{\link{read_fb_files}}
 }
