+++ man/read_fb_files.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/read_and_format_files.R
-\name{read_all_files}
-\alias{read_all_files}
+\name{read_fb_files}
+\alias{read_fb_files}
 \title{Read all the .csv files in a folder and format the data.}
 \usage{
-read_all_files(path_to_csvs, colnames, coltypes)
+read_fb_files(path_to_csvs, colnames, coltypes)
 }
 \arguments{
 \item{path_to_csvs}{Path to the folder where the .csv files are stored}
@@ -28,7 +28,7 @@ specific formatting to columns across these files."
 \examples{
 
 
-# read_all_files(path_to_csvs = 'data/',
+# read_fb_files(path_to_csvs = 'data/',
 #                 colnames = c("lat", "lon", 
 #                              "quadkey", "date_time", 
 #                              "n_crisis", "percent_change"),
@@ -42,7 +42,7 @@ specific formatting to columns across these files."
 
 }
 \seealso{
-\code{\link{format_data}}
+\code{\link{format_fb_data}}
 
 \code{\link[readr]{read_csv}}
 }
