+++ NAMESPACE
@@ -7,7 +7,7 @@ export(create_qk_grid)
 export(create_raster)
 export(extract_qk_coord)
 export(extract_tile_coord)
-export(format_data)
+export(format_fb_data)
 export(grid_to_polygon)
 export(ground_res)
 export(latlong_to_pixelXY)
@@ -21,7 +21,7 @@ export(polygon_to_raster)
 export(qkmap_app)
 export(quadkey_to_latlong)
 export(quadkey_to_tileXY)
-export(read_all_files)
+export(read_fb_files)
 export(regular_qk_grid)
 export(tileXY_to_pixelXY)
 export(tileXY_to_quadkey)
