### 2024-01-13 1d7b518
Improvements in functions and roxygen docs as requested in (ropensci/software-review#619)
