+++ R/read_and_format_files.R
@@ -20,35 +20,39 @@
 #'
 #' @examples
 #'
-#' 
-#' # read_fb_files(path_to_csvs = 'data/',
-#' #                 colnames = c("lat", "lon", 
-#' #                              "quadkey", "date_time", 
-#' #                              "n_crisis", "percent_change"),
-#' #                 coltypes = list(
-#'  #                                lat = 'd',
-#'  #                                lon = 'd',
-#'  #                                quadkey = 'd',
-#'  #                                date_time = 'T',
-#'  #                               n_crisis = 'c',
-#'  #                                percent_change = 'c'))
+#' \dontrun{
+#'  read_fb_mobility_files(path_to_csvs = 'data/',
+#'                  colnames = c("lat", "lon", 
+#'                               "quadkey", "date_time", 
+#'                               "n_crisis", "percent_change"),
+#'                  coltypes = list(
+#'                                 lat = 'd',
+#'                                 lon = 'd',
+#'                                 quadkey = 'd',
+#'                                 date_time = 'T',
+#'                                 n_crisis = 'c',
+#'                                 percent_change = 'c'))
+#' }
 #'
-read_fb_files <- function(path_to_csvs, 
+read_fb_mobility_files <- function(path_to_csvs, 
                            colnames, 
                            coltypes){
-
-  
 
   # This data always have the same format
   fnames <- list.files(path = path_to_csvs,
                        pattern = "\\.csv$")
+  
+  # Give an error if there are not empty files
+  if (length(fnames) == 0) {
+    stop("Error:.csv files not found in the specified directory.")
+  }
 
   fnames <- paste0(path_to_csvs,
                    fnames)
   
   # There could be empty files, let's remove them
   fnames <- fnames[file.info(fnames)$size != 0]
-
+  
   data <- purrr::map_dfr(fnames,
                      readr::read_csv,
                      col_select = dplyr::all_of(colnames), # tidyselect
@@ -62,10 +66,7 @@ read_fb_files <- function(path_to_csvs,
     message(paste("The files with the following combinations of",
                   "days and times are not present or have 0KB"))
     print(missing_combinations(data))
-  }else{
-    message("There aren't missing dates or times")
   }
-
   data
 
 }
@@ -73,8 +74,8 @@ read_fb_files <- function(path_to_csvs,
 
 #' Format the data
 #'
-#' @description This function modifies the format of three columns the provided
-#' data.
+#' @description This function removes unnecessary characters such as `\\N`
+#' and ensures that the format of the date and QuadKeys is correct.
 #'
 #' @param data A dataframe with a quadkey, date_time, country columns and 
 #' other numeric variables
@@ -82,13 +83,14 @@ read_fb_files <- function(path_to_csvs,
 #' @return A dataframe.
 #' @export
 #'
-#' @seealso \code{\link{read_fb_files}}
+#' @seealso \code{\link{read_fb_mobility_files}}
 #'
 #' @examples
 #'
-#' #data(onefile)
-#' #format_fb_data(data = onefile)
-#'
+#'\dontrun{
+#' fb_mobility_file <-  read.csv("cityA_2020-04-26_0.csv")
+#' format_fb_data(data = fb_mobility_file)
+#'}
 format_fb_data <- function(data){
 
   # remove scientific notation
@@ -98,7 +100,7 @@ format_fb_data <- function(data){
   # change date format
   data$day <- lubridate::date(data$date_time)
 
-  data$time <- as.numeric(format(as.POSIXct(data$date_time, 
+  data$hour <- as.numeric(format(as.POSIXct(data$date_time, 
                                             format = "%Y-%m-%d %H%M"), 
[…]
