### 2024-01-13 c2de42e
Removes duplicated LICENSE file (ropensci/software-review#619)
