### 2024-01-17 5749abc
Add a vignette for qkmap app and the # of QuadKeys in the app UI + minor UI changes (ropensci/software-review#619)
