+++ inst/app/app.R
@@ -1,6 +1,6 @@
-qkmap_app <- function(...) {
+qkmap_app <- function(...){
   ui <- bslib::page_navbar(
-    title = "QuadKey Map Visualizer",
+    title = "QuadKey Visualizer App",
     bg = "#ffffff",
     theme = bslib::bs_theme(version = 5,
                             bootswatch = "simplex",
@@ -8,184 +8,177 @@ qkmap_app <- function(...) {
                                                            wght = "400"),
                             heading_font = bslib::font_google("Raleway",
                                                               wght = "200")),
-     bslib::nav_panel(title = 'Quadkey',
-              bslib::layout_sidebar(class = 'p-0',
-                     sidebar = bslib::sidebar(width = 300,
-        shiny::textInput('qk',
-                  'Visualize the upper-left coordinate of the QuadKey',
-                  placeholder = "complete and click search"),
-        shiny::actionButton('search', 'Search'),
-        shiny::hr(),
-        DT::DTOutput('qkvalues')
-
-      ),
-      leaflet::leafletOutput('mapqk'))
-
+    bslib::nav_panel(title = 'Quadkey',
+                     bslib::layout_sidebar(class = 'p-0',
+                                           sidebar = bslib::sidebar(width = 300,
+                                                                    shiny::textInput('qk',
+                                                                                     'Visualize the upper-left coordinate of the QuadKey',
+                                                                                     placeholder = "complete and click search"),
+                                                                    shiny::actionButton('search', 'Search'),
+                                                                    shiny::hr(),
+                                                                    DT::DTOutput('qkvalues')),
+                                           leaflet::leafletOutput('mapqk'))
     ),
     bslib::nav_panel(title = 'Grid',
-           bslib::layout_sidebar(class = 'p-0',
-                     sidebar = bslib::sidebar(
-                       shiny::p("Write the coordinates in decimal degrees"),
-
-            bslib::layout_columns(
-          shiny::textInput('xmin', 'xmin'),
-          shiny::textInput('ymin', 'ymin')
-        ),
-        bslib::layout_columns(
-          shiny::textInput('xmax', 'xmax'),
-          shiny::textInput('ymax', 'ymax')
-        ),
-        shiny::sliderInput('levelofdetail',
-                    'Select the level of detail:',
-                    min = 1,
-                    max = 23,
-                    value = 6),
-        shiny::actionButton('grid', 'Create Grid'),
-        shiny::hr(),
-        shiny::p(paste("Note that for higher levels of detail,",
-                       "smaller areas are preferable to prevent",
-                       "long processing times."))
-
-      ),
-      leaflet::leafletOutput('mapgrid')))
-
-
+                     bslib::layout_sidebar(class = 'p-0',
+                                           sidebar = bslib::sidebar(
+                                             shiny::p("Write the coordinates in decimal degrees"),
+                                             
+                                             bslib::layout_columns(
+                                               shiny::textInput('xmin', 'xmin'),
+                                               shiny::textInput('ymin', 'ymin')
+                                             ),
+                                             bslib::layout_columns(
+                                               shiny::textInput('xmax', 'xmax'),
+                                               shiny::textInput('ymax', 'ymax')
+                                             ),
+                                             shiny::sliderInput('levelofdetail',
+                                                                'Select the level of detail:',
+                                                                min = 1,
+                                                                max = 23,
+                                                                value = 6),
+                                             shiny::actionButton('grid',
+                                                                 'Create Grid'),
+                                             shiny::hr(),
+                                             textOutput("qk_nr"),
+                                             shiny::hr(),
[…]
