### 2024-01-17 78bbcf0
Removes loop from create_qk_grid (ropensci/software-review#619)
