### 2024-01-19 5adbf3c
Fix check() issues - (ropensci/software-review#619) - Remove <> from BugReports: in the DESCRIPTION file - Remove .md from LICENSE - Update shinytests2 after changing shiny UI in previous commit - grid_to_polygon dplyr::filter() changes to base R to avoid issues - Fix qk_to_latlong_function to include a column with the level of detail - extract_*_coords() functions create columns needed to avoid warning - Improvement in rowwise() implementation. Tidyselect.
