+++ R/extract_coords.R
@@ -31,6 +31,14 @@ extract_qk_coord <- function(data){
   if (!"quadkey" %in% colnames(data)) {
     stop("Please ensure that the dataset contains a column named 'quadkey'.")
   }
+  
+  # initialize columns to avoid warnings
+  data$qk_tileX <- NA
+  data$qk_tileY <- NA
+  data$tl_pxx <- NA
+  data$tl_pxy <- NA
+  data$pxy_lat <- NA
+  data$pxy_lon <- NA
 
   for(i in seq_len(nrow(data))){
     # check that the data has the correct dimensions for this analysis
@@ -101,6 +109,13 @@ extract_tile_coord <- function(data, level){
                "and 'tileY'"))
   }
 
+  
+  # initialize columns to avoid warnings
+  data$tl_pxx <- NA
+  data$tl_pxy <- NA
+  data$pxy_lat <- NA
+  data$pxy_lon <- NA
+  
   for(i in seq_len(nrow(data))){
    # conversion from tile coordinates to geographic coordinates
     ttop <- tileXY_to_pixelXY(tileX = data$tileX[i],
