+++ DESCRIPTION
@@ -1,4 +1,4 @@
-Package: Quadkeyr
+Package: quadkeyr
 Title: Tools for converting QuadKeys used in Microsoft's Bing Maps Tile System into raster images
 Version: 0.0.0.9000
 Authors@R: 
@@ -10,7 +10,7 @@ Authors@R:
 Description: 'Quadkeyr' enables the generation of raster images based on QuadKeys, facilitating efficient integration of Bing Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
 License: MIT + file LICENSE
 URL: <https://fernandez-lab-wsu.github.io/quadkeyr/>, <https://github.com/Fernandez-Lab-WSU/quadkeyr>
-BugReports: <https://github.com/Fernandez-Lab-WSU/quadkeyr/issues>
+BugReports: https://github.com/Fernandez-Lab-WSU/quadkeyr/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
