### 2024-01-20 e48b9ff
Update R dependency version in DESCRIPTION to 4.0.1 (ropensci/software-review#619)
