+++ DESCRIPTION
@@ -28,7 +28,7 @@ Suggests:
     viridis (>= 0.6.4)
 VignetteBuilder: knitr
 Depends: 
-    R (>= 2.10)
+    R (>= 4.0.1)
 LazyData: true
 Imports: 
     dplyr (>= 1.1.2),
