### 2024-01-20 183788c
Change time variable name for hour in data and functions that use it (ropensci/software-review#619)
