+++ R/data.R
@@ -10,11 +10,11 @@
 #' }
 "data_provided"
 
-#' result_read_all_files: Fake dataset
+#' result_read_fb_mobility_data: Fake dataset
 #'
 #' A data frame similar to a the potential output of read_all_files().
 #' 
-#' @format ## `result_read_all_files`
+#' @format ## `result_read_fb_mobility_data`
 #' A data frame with 134,492 rows and 9 columns:
 #' \describe{
 #'   \item{lat}{Latitude of QuadKey centroid}
@@ -25,6 +25,6 @@
 #'   \item{n_crisis}{Variable}
 #'   \item{percent_change}{Variable}
 #'   \item{day}{Day in format %Y-%m-%d}
-#'   \item{time}{Hour of the day as a number between 1 and 24}
+#'   \item{hour}{Hour of the day as a number between 1 and 24}
 #'   }
-"result_read_all_files"
+"result_read_fb_mobility_data"
