### 2024-01-20 #619 @flor14
@vincentvanhees and @emilyriederer

I have been reviewing @vincentvanhees comments, and here are the actions I have taken to address most of them. There is pending section at the end of this issue, as I think that the final details should be addressed after Maria Paula's review to avoid overworking.

The revision has been quite detailed, so I want to express my gratitude again for all the comments provided.

## 1. Vignettes and README:
- I renamed the vignettes keeping the names you suggested, except for `create_your_own_quadkeys.Rmd`, which I changed to `from_quadkey_data_to_raster.Rmd`. The only difference between the second and third vignette is whether the data comes from Facebook or not, but the process is quite similar otherwise. Fernandez-Lab-WSU/quadkeyr@c34a2a61.
- The vignettes filenames are organized alphabetically (I tried to number them, but numbers in filenames made `check()` fail so I replace numbers with letters). Those changes are here: Fernandez-Lab-WSU/quadkeyr@9c81a05
- I added a call to README in the vignettes. I linked the `pkgdown` site's first page as the README, as it seems neater. You can see the changes here:Fernandez-Lab-WSU/quadkeyr@662bb73

####  Specific comments on`get_grid_from_quadkeys.Rmd` (now called `B-from_quadkey_data_to_raster.Rmd`):

-  Improved the text, focusing on explaining `tileX` and `tileY` in the first vignette. I will re-read it at the end of the revision again but the changes are here: Fernandez-Lab-WSU/quadkeyr@b3f1fbc
- Fixed the autonumbering issue manually in the vignettes. Modifying the `yaml` as you suggested never worked, I suspect that it has to do with the fact that [a vignette is a simplified RMarkdown file](https://bookdown.org/yihui/rmarkdown/r-package-vignette.html) with fewer functionalities.
The documentation was improved and more images were added Fernandez-Lab-WSU/quadkeyr@b3f1fbc

#### specific comments on `create_rasters_from_grid.Rmd` (now called `C-from_facebook_quadkey_data_to_raster.Rmd`)
- Replace `remotes` in the `README`. Fernandez-Lab-WSU/quadkeyr@6716486

#### specific comments `quadkey_conversion.Rmd` (now called `A-bing_map_tile_system_functions.Rmd`):
- Text improvement. Fernandez-Lab-WSU/quadkeyr@b3f1fbc

## 2. R functions:
- All these were small comments regarding the functions that I have solved in Fernandez-Lab-WSU/quadkeyr@1d7b518 and other minor ones in  Fernandez-Lab-WSU/quadkeyr@183788c
- The constants/numbers in the functions are the ones used in the [official documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system). They define the mathematical relationship among the variables and coordinates used, proposing functions. I am only making these functions available in R (and adding the conversion to raster). I agree that adding comments about this could be helpful for the user. I wrote a note in the pending section to double-check that I have added all the comments for the final review.

## 3. Test function:
- Also addressed in Fernandez-Lab-WSU/quadkeyr@1d7b518

## 4. Shiny app:
- To guide the user better about potential performance issues, the app now displays the number of QuadKeys, as you can see in this image: ![image](https://github.com/ropensci/software-review/assets/29610858/2e531600-5c70-4bf3-882d-9e9d8201cea7)
- Also, I have added a specific vignette for the app functionality only: Fernandez-Lab-WSU/quadkeyr@5749abc.
- In case it wasn't noticed, there has always been a [notification](https://github.com/Fernandez-Lab-WSU/quadkeyr/blob/45340af79b1d5def553292755b37198bd2205681/inst/app/app.R#L152) regarding the size of the grid and standard computing times for the user.

- I have read about the use of the `inst/` folder for Shiny apps in packages for example in this [R community thread](https://community.rstudio.com/t/shiny-app-as-a-package-directory-structure-and-www-folder/135323) and this [blogspot](https://www.r-bloggers.com/2021/02/framework-for-shiny-apps-in-r-packages/). I understand that locating the Shiny app in `inst/` is a more organized way of structure a package when it has multiple functions and not a standalone app.
Also, when you create the tests with `shinytest2` in a package structure the documentation guides you in [both solutions](https://rstudio.github.io/shinytest2/articles/use-package.html), using the `inst/` folder or not.

## 5. General / Other comments:
- Fixed LICENSE issue Fernandez-Lab-WSU/quadkeyr@c2de42e
- Fixed R dependency version in DESCRIPTION file Fernandez-Lab-WSU/quadkeyr@e48b9ff
This one called my attention I have create the DESCRIPTION file using the `usethis` functions, so 2.10 was selected automatically.

## 6. Suggestions (no need to respond):
See pending actions.

## Pending Actions (To be addressed at the end of the revision):

- [ ] README: Reconsider the clarity of the section "What can this package do for you?" after going through my comments on the vignettes above.
- [ ] Reduce the loops and the empty lines. Check identation in all the files and confirm that all the constants have comments explaining where they came from.
- [ ] Re-read vignettes text.
- [ ] Check that `pkgcheck` is passing

Feel free to reach out if you have any questions or need further clarification.
