### 2024-01-20 #619 @vincentvanhees
Thanks @flor14. I will have a closer look when you have also received and processed comments from the María.
Regarding the auto-numbering: I have not had this problem myself and doubt vignettes as such prohibit this. See ([link to example of my vignettes](https://github.com/wadpac/GGIR/blob/master/vignettes/GGIR.Rmd#L6)), maybe you could have a look whether I am doing something very different than what you have been trying.
