### 2024-01-23 #619 @mpaulacaldas
@flor14 @emilyriederer @vincentvanhees  - With apologies for the delay, I leave below my review. As you will see, it is extensive, so please feel free to ignore the more minor suggestions. I'm happy to discuss any points that are unclear or help brainstorm solutions, so please don't hesitate!

---

  Florencia and I worked together on the Spanish translation of a chapter in R for Data Science in 2018. That collaboration went smoothly, but we have not had the chance to work together since.

## Review Comments

Thank you very much for the opportunity to review this package. As someone who has struggled with quadkeys in the past, I think this package addresses a very real gap in the R spatial software space, and I believe it will be useful to many social scientists.

I leave below detailed comments on the different aspects of the package. The review is quite extensive, so please don’t hesitate to ignore the more minor suggestions. Likewise, I’m happy to discuss any points that are unclear or help brainstorm solutions.

This is an excellent README. If I could provide only one minor comment, it would be to put the section “What can this package do for you?” *before* “What are quadkeys and tile maps?”

### Vignettes

Overall, I find the vignettes are informative and easy to understand, with truly excellent pedagogical figures.

I leave below my comments for possible improvements, to complement some of the points raised previously by Vincent:

-   On **`C-from_facebook_quadkey_data_to_raster.Rmd`**:
    -   The build of this vignette is failing as of https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/183788c0eb24d47b5da06fe176680db27e4da886 (the commit I am using for this review). This is because the data frame passed to `apply_weekly_lag()` does not contain an `hour` column, that it needs for it to run.
    -   As I understand it, the objective of this vignette is to showcase *how to create a raster from a table where cells are identified with quadkeys*. To do so, you use Facebook mobility data as an example, but the logic can be applied to other quadkey-identified datasets. With this in mind, I think it would be useful to have a more general title, for example: “Converting a quadkey-identified data set to a raster”. In the first sentences of the vignette, you can mention that your example will use FB mobility data, but the same logic could be applied to any data set with column with quadkey-identifiers.
    -   The example in the vignette is not reproducible, as it depends on the files in `"../geocovid/data/rasters/city/"` – a relative path only available in the developer’s computer. I see two ways to fix this issue:
        -   If the data is small enough, add it to the package under `inst` and access it using `system.file()` or `fs::path_package()`, as suggested in the [data chapter](https://r-pkgs.org/data.html#sec-data-system-file) of the R packages book.
        -   If the data is too large, create a helper that *downloads* the files into a directory and reads them from there (FB mobility data may be found under a stable link).
    -   In step 2 and 3, it would be useful to explicitly mention in the text that the number of quadkeys went from 150 to 396 (though the info is available in the printed `sf` object, it is a bit difficult to parse). Explicitly mentioning these numbers may make it easier for readers to understand what you mean by “completing the grid”.
    -   Seeing step 4, it is unclear to me why we bother expanding the grid, if in the end the `inner_join()` will get rid of all the quadkeys from `polygrid` that are implicitly missing from `files`. Perhaps it would be useful to re-instate that the grid expansion is performed because creating the polygons (and later the rasters) in this implementation requires a full grid.
    -   The resolution of the last image is off in my display using the RStudio viewer.

-   On **`B-from_facebook_quadkey_data_to_raster.Rmd`**:
    -   I think the name/title of this vignette could still be improved. I would suggest “Building a quadkey grid from a bounding box” (e.g. and a file name like `quadkey_grid_from_bbox.Rmd` for example) or something along those lines. In the end, the main difference between this vignette and `C-from_facebook_quadkey_data_to_raster` is that you are showing how to build your own quadkey grid over a given extent, but all other steps (i.e. going from a data set with quadkeys to a raster) are the same. I imagine most users will end up having a workflow like `C-from_facebook_quadkey_data_to_raster`, but only few will need to build their own quadkey grid.
    -   I assume `xmin`, `xmax`, `ymin` and `ymax` refer to the coordinates of the *bounding box*, following the common notation from `sf`. If so, I would advise using the term *bounding box* explicitly, since it is a term that most spatial R users will be familiar with.
    -   In this vignette and throughout the package in general, it would be better to use the word *zoom level* explicitly instead of “level of detail”. Zoom level is the term that package users are most likely to look for when they are skimming the documentation, and it is also the parameter name used in other packages (e.g. `slippymath`)

-   On **`A-bing_map_tile_system_functions.Rmd`**:
    -   This is a minor note, but I think it’s easier (and still correct) to say “quadkey” instead of “quadkey number”.
    -   Bravo for including the caveat in Section 3. Perhaps a more eye-catching title would help users understand the importance of reading this section (e.g. “Caveats when converting coordinates”)

-   On **`D-quadkey_visualization_app.Rmd`**:
    -   It’s a really nice and valuable addition to have a vignette introducing the Shiny app, with screenshots showing common behaviours. I will leave here some minor improvements I think you could make to the *app*:
        -   I think it would be useful if the *Quadkey* tab, showed the boundaries of the user-specified quadkey in addition to only the upper-left corner. As an analyst, I often want to visualise the *area* covered by a specific quadkey more than the coordinates of its extreme points. If you introduce these changes, I think a better name for the tab could be “Visualising a quadkey”.
        -   I think the *Grid* tab could have a better name, so users know that its goal is to produce a quadkey grid from a set of coordinates and a zoom level. I would suggest “Creating a quadkey grid”.

-   Lastly, and going slightly against @vincentvanhees suggestion (sorry!) I think you should *remove* the alphabetical tags from the vignettes names. This is because locally, users access vignettes using `vignette("file_name_without_extension")`, and longer file names can be harder to type and remember. As someone having working with quadkeys in the past, the most useful vignette seems to be `C-from_facebook_quadkey_data_to_raster`, so if you can, I would prioritise making that specific name shorter, while still informative (not an easy task at all!)

### Function documentation

-   Following on the above, it would be useful to add to the function documentation of `apply_weekly_lag()`, a line in the Description clarifying that these are specific to the format of FB mobility data.

-   Is `quadkeyr::clip()` supposed to be exported? Its documentation mentions that this is an internal function and doesn’t provide much detail on functionality. It also masks `graphics::clip()`, so if you don’t think it will be much use to end-users, I would suggest to keep it internal.If you want to roxygen to produce a documentation file for the function, while preventing the function from being exported, you can use the `@keywords internal` tag instead of `@export`.

-   Similar comment for `complete_grid_for_polygons()`. Is this function meant to be exported? Are you exporting it because it may be useful for advanced users? If so, it would be useful to mention it in the function documentation.

-   The code style is at times inconsistent across the documentation (examples and vignettes). For example, `=` and `<-` both being used for assignment, sometimes even within the same example (e.g. `create_raster()`). Using `styler::style_pkg()` to catch these may be helpful.

- What is the difference between `extract_qk_coord()` and `extract_tile_coord()`? It is unclear from the documentation and the examples. Doing a quick test, I also get that both functions return the same result, with the exception that `extract_qk_coord()` does not require a `level` argument.
   ```r
    grid <- create_qk_grid(
      xmin = -59,
      xmax = -40,
      ymin = -38,
      ymax = -20,
      level = 6
    )

    tile_result <- extract_tile_coord(data = grid$data, level = 6)
    qk_result   <- extract_qk_coord(data = grid$data)

    waldo::compare(tile_result, qk_result)
    #> ✔ No differences
   ```

-   The example in `format_fb_data()` is not reproducible for users and is wrapped in a `dontrun{}`. Couldn’t you provide an internal data frame mimicking the format of facebook mobility data? The name `fb_mobility_file` that you use in the function documentation sounds like a great name for this example data frame. Alternatively, you could also demo how to download the file from a stable link with `download.file()`, saving to a temporary directory, and reading from there.

-   In `ground_res()`, it would be useful to add a link to the Microsoft documentation. Same goes for `mapscale()`, `mapsize()`.

-   In `latlong_to_quadkey()`, do the *returned* coordinates correspond to the upper-left corner of the pixel? Is this where the caveat from the vignette applies? If so, it wouldn’t hurt to repeat it again in the function documentation. Same comment goes for `pixelXY_to_latlong()`, `tileXY_to_pixelXY()`

-   It’s a shame that the example of `polygon_to_raster()` is not runnable nor reproducible, as this is the example that best describes that the entire workflow. My comments and suggestions to fix it are the same as mentioned above for the `C-from_facebook_quadkey_data_to_raster.Rmd` vignette. These would also apply to the function documentation of `read_fb_mobility_fiels()`.

-   There is some inconsistency in the argument names for `quadkey_to_tileXY()` and `quadkey_to_langlot()`, despite both taking a character quadkey as primary argument. Some suggestions:

    -   Making both functions vectorised, so they can each accept a quadkey character vector of any length, OR…
    -   Keep the current behaviour, but change the argument name of `qk` to `quadkey` in `quadkey_to_tileXY()`, for more consistency with `quadkey_to_latlong()`

-   The documentation of `quadkey_to_tileXY` says the argument needs to be a character vector, but in the example, you use a numeric vector and rely on type coercion. I suggest fixing the example to use an argument, **but also** making `quadkey_to_tileXY` fail explicitly if the input type is not character.

-   As mentioned above, `apply_weekly_lag()` is currently failing for the `C-from_facebook_quadkey_data_to_raster.Rmd` vignette. When you implement a fix, it would perhaps be a good idea to add a unit test checking that the function works for the data used in the vignette.

-   The example provided in `complete_grid_polygons()` returns a data frame where the quadkey column is filled with `NA`, which does not seem like the intended behaviour. If you fix this, I suggest adding a unit-test based on the example. It should be easy since your example is already in a compact-enough format.

-   When working though the common use-case of creating a raster from a vector of quadkeys (see reprex below), I noticed two potential improvements for `grid_to_polygon()` and `regular_qk_grid()`:

    -   `grid_to_polygon()` fails unless the input data set contains the variables `tileX`, `tileY`, `quadkey` and `geometry` in that *specific order*. I would suggest mentioning this in the function documentation and to provide a more informative error message.
    -   `regular_qk_grid()` should warn when the provided grid is already regular, instead of throwing an error. This would make piping `quadkeyr` functions easier for users.

    <summary>An example of a common workflow</summary>

     ```r
    library(quadkeyr)
    #>
    #> Attaching package: 'quadkeyr'
    #> The following object is masked from 'package:graphics':
    #>
    #>     clip
    library(ggplot2)

    medellin_qks <- c(
      "032230320233",
      "032230320322",
      "032230322011",
      "032230322100",
      "032230322013",
      "032230322102"
      )

    (medellin_coords <- quadkey_to_latlong(medellin_qks))
    #> Simple feature collection with 6 features and 1 field
    #> Geometry type: POINT
    #> Dimension:     XY
    #> Bounding box:  xmin: -75.67383 ymin: 6.227934 xmax: -75.58594 ymax: 6.402648
    #> Geodetic CRS:  WGS 84
    #>        quadkey                   geometry
    #> 6 032230322102 POINT (-75.58594 6.227934)
    #> 5 032230322013 POINT (-75.67383 6.227934)
    #> 4 032230322100 POINT (-75.58594 6.315299)
    #> 3 032230322011 POINT (-75.67383 6.315299)
    #> 2 032230320322 POINT (-75.58594 6.402648)
    #> 1 032230320233 POINT (-75.67383 6.402648)

    # My grid is already regular, so in this case there is no need to regularise.
    # However, I would suggest changing this default behaviour so it throws a
    # warning instead of an error, and silently returns the originally provided
    # grid. This way the function can be used more easily in pipes.
    regular_qk_grid(medellin_coords)
    #> Error in regular_qk_grid(medellin_coords): The grid is already complete, this function is not necessary

    # Why does grid_to_polygon() fail if my grid is already regular + it has the
    # quadkeys and coordinates? I imagine this is because grid_to_polygon() expects
    # the tileX and tileY columns returned by extract_qk_coord(). If so, it would
    # be useful to mention it in the function documentation, and throw an error
    # saying that columns are missing
    grid_to_polygon(medellin_coords)
    #> Warning in max(data$tileX): no non-missing arguments to max; returning -Inf
    #> Warning in max(data$tileY): no non-missing arguments to max; returning -Inf
    #> Warning in min(data$tileX): no non-missing arguments to min; returning Inf
    #> Error in seq.default(min(data$tileX), textX): 'from' must be a finite number

    # My workaround to add the columns
    medellin_tileXY <- medellin_coords$quadkey |>
      purrr::set_names() |>
      purrr::map_dfr(quadkey_to_tileXY, .id = "quadkey")
    medellin_tileXY
    #> # A tibble: 6 × 4
    #>   quadkey      tileX tileY level
    #>   <chr>        <dbl> <dbl> <int>
    #> 1 032230322102  1188  1977    12
    #> 2 032230322013  1187  1977    12
    #> 3 032230322100  1188  1976    12
    #> 4 032230322011  1187  1976    12
    #> 5 032230320322  1188  1975    12
    #> 6 032230320233  1187  1975    12

    medellin_polygons <- medellin_coords |>
      dplyr::left_join(medellin_tileXY, by = "quadkey") |>
      # grid_to_polygon() would fail if these columns were not in this exact order.
      # Ideally, it would be great if grid_to_polygon() were not as strict when it
      # comes to the order of the columns. If tricky to implement, it would be
      # useful if the restriction were at least mentioned in the documentation
      dplyr::select(tileX, tileY, quadkey, geometry) |>
      grid_to_polygon()

    ggplot(medellin_polygons) +
      geom_sf() +
      geom_sf_text(aes(label = quadkey)) +
      theme_void()
    #> Warning in st_point_on_surface.sfc(sf::st_zm(x)): st_point_on_surface may not
    #> give correct results for longitude/latitude data
    ```

### API design

These comments touch more on the API design choices, which fall a bit out-of-scope of the reviewing guidelines from rOpenSci. I raise them here in case you find them useful, but I completely understand if you prefer not to implement them, since they can be time-consuming :smile:

-   From the user perspective, I find it a bit difficult to differentiate between the functions that are specific to the FB mobility data (e.g. `apply_weekly_lag()`, `missing_combinations()`) and those that are more general and could be used in other applications (e.g. `create_quadkey_grid()`). Perhaps a way to remove some of this confusion could be to create different sections in the pkgdown website, bundling together functions that are specific to FB mobility data and those that are more general. usethis uses a similar approach, see for example their [function reference](https://usethis.r-lib.org/reference/index.html#create) and how they define the categories via the [`_pkgdown.yml`](https://github.com/r-lib/usethis/blob/5986ab68d14454c78013082e36f9ec2bcd216b11/_pkgdown.yml#L37-L44). If you want to take it a step further, it could also be useful to use a common prefix for FB-specific operations (e.g. `fb_apply_weekly_lag()`, `fb_format_data()`).

-   I find the name of `create_raster()` confusing. As a spatial R user, the name `create_raster()` leads me to believe that the function returns a `raster` `Raster*` object, while in reality it returns a `stars` object. In my opinion, a slightly different name (e.g. `create_stars()` or
    `create_stars_raster()`) would be slightly less confusing.

-   Similarly, when I read `extract_qk_coord()` or `extract_tile_coord()`, I expect a function that behaves similarly to `terra::extract()`/`raster::extract()` (i.e. which extract the value of a raster based on a polygon or point; functions that take a raster and a polygon/point object as arguments). However, `extract_qk_coord()` and `extract_tile_coord()` are *not* doing spatial extractions. Perhaps a name like `get_qk_coord()` would be better.
