### 2024-01-23 #619 @flor14
Thank you @mpaulacaldas for your review!

All the comments are relevant, and I have gained new insights after using the package myself. I will probably include some minor changes on my side during the revision. I will open issues so everything is clear.

My goal is to finish this by the end of January.

I am really happy to have such good reviewers.
