### 2024-01-25 f914c51
Delete data example file with old name (ropensci/software-review#619)
