### 2024-01-25 feb1cc0
Change text order in README as described in ropensci/software-review#619
