+++ README.Rmd
@@ -22,6 +22,15 @@ knitr::opts_chunk$set(
 ⚠️ This package is under review
 
 
+### What can this package do for you?
+The `quadkeyr` R package presents a comprehensive toolkit tailored for
+generating raster images from data reported by QuadKeys within [Microsoft's Bing Maps Tile System](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system). 
+Designed to integrate Bing Maps data into R workflows,
+this package facilitates the creation of QuadKey grids and raster images
+and introduces specialized functions for
+the processing of [Facebook Mobility data](https://dataforgood.facebook.com/).
+
+
 ### What are QuadKeys in Tile Maps?
 
 Tile maps divide the Earth's surface into a grid of tiles, with each tile
@@ -36,14 +45,6 @@ mapping applications such as [Microsoft Bing Maps](https://learn.microsoft.com/e
 ```{r qk, echo = FALSE, out.width= "90%", fig.align='center', fig.cap= "The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage."}
 knitr::include_graphics("man/figures/quadkeys.jpg")
 ```
-
-### What can this package do for you?
-The `quadkeyr` R package presents a comprehensive toolkit tailored for
-generating raster images from data reported by QuadKeys within [Microsoft's Bing Maps Tile System](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system). 
-Designed to integrate Bing Maps data into R workflows,
-this package facilitates the creation of QuadKey grids and raster images
-and introduces specialized functions for
-the processing of [Facebook Mobility data](https://dataforgood.facebook.com/).
 
 
 ---
