### 2024-01-25 1f0d917
Change vignette title C-from_facebook_quadkey_data_to_raster.Rmd and other suggestions in ropensci/software-review#619
