### 2024-01-25 6e00388
Improve explanation of what is complete grid in C-from_quadkey_identified_facebook_mobility_data_to_rasters (ropensci/software-review#619)
