### 2024-01-25 c8f9728
Make clip() internal and update NAMESPACE (ropensci/software-review#619)
