+++ man/clip.Rd
@@ -29,3 +29,4 @@ documentation.
      max_value = 5)
 
 }
+\keyword{internal}
