+++ man/result_read_fb_mobility_data.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{result_read_all_files}
-\alias{result_read_all_files}
-\title{result_read_all_files: Fake dataset}
+\name{result_read_fb_mobility_data}
+\alias{result_read_fb_mobility_data}
+\title{result_read_fb_mobility_data: Fake dataset}
 \format{
-\subsection{\code{result_read_all_files}}{
+\subsection{\code{result_read_fb_mobility_data}}{
 
 A data frame with 134,492 rows and 9 columns:
 \describe{
@@ -17,12 +17,12 @@ A data frame with 134,492 rows and 9 columns:
 \item{n_crisis}{Variable}
 \item{percent_change}{Variable}
 \item{day}{Day in format \%Y-\%m-\%d}
-\item{time}{Hour of the day as a number between 1 and 24}
+\item{hour}{Hour of the day as a number between 1 and 24}
 }
 }
 }
 \usage{
-result_read_all_files
+result_read_fb_mobility_data
 }
 \description{
 A data frame similar to a the potential output of read_all_files().
