+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(apply_weekly_lag)
-export(clip)
 export(complete_grid_for_polygons)
 export(create_qk_grid)
 export(create_raster)
