### 2024-01-28 6054769
Simplify code in extract_qk_coord and rename variables in extract_tile_coord (ropensci/software-review#619)
