### 2024-01-28 2772dd9
Include link to the official documentation for selected functions. Update roxygen docs (ropensci/software-review#619)
