+++ R/grid_to_polygon.R
@@ -65,8 +65,8 @@ complete_grid_for_polygons <- function(data){
 
 #' Convert a grid of QuadKeys to square polygons
 #'
-#' @description The main argument of this function, the grid of QuadKeys points
-#' representing lat/long WG84 coordinates specifically indicate the upper-left 
+#' @description The main argument of this function, the grid of 
+#' geographic coordinates (lat/long WG84) represents the upper-left 
 #' corner of the QuadKey.
 #' To transform these coordinates into square polygons, the function
 #' supplements the grid by adding a row and column of tiles. These points
