+++ man/pixelXY_to_latlong.Rd
@@ -30,3 +30,7 @@ pixelXY_to_latlong(pixelX = 768,
 
 
 }
+\seealso{
+\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
+Microsoft Bing Maps Tile System documentation}
+}
