+++ man/quadkey_to_tileXY.Rd
@@ -21,3 +21,7 @@ Microsoft Bing Maps Tile System documentation.
 
 quadkey_to_tileXY(qk = 213)
 }
+\seealso{
+\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
+Microsoft Bing Maps Tile System documentation}
+}
