### 2024-01-28 f5966b6
Replace extract_* with get_* (ropensci/software-review#619) - Replace the functions changed in tests, examples and other functions. - Solve multiple small updates after check(), mostly in relation with updates in names of data and functions. - Change R version to 4.0 to be less strict with the dependencies. - Delete old copy of result_read_all_files that has been replaced.
