+++ R/create_raster.R
@@ -24,7 +24,7 @@
 #'                      ymax = -34,
 #'                      level = 12)
 #'
-#'  grid_coords <- extract_qk_coord(data = grid$data)
+#'  grid_coords <- get_qk_coord(data = grid$data)
 #'
 #'  polygrid <- grid_to_polygon(grid_coords)
 #'
