+++ tests/testthat/test-get_coords.R
@@ -7,13 +7,13 @@ grid <- create_qk_grid(
   ymax = -34,
   level = 12)
 
-# QuadKeys - extract_qk_coord()
+# QuadKeys - get_qk_coord()
 test_that("Check existence of 'quadkey' column", {
   # Create a dataframe without 'quadkey' column
   df_missing_column <- data.frame(other_column = c(1, 2, 3))
   
   # Test case where 'quadkey' column is missing
-  expect_error(extract_qk_coord(df_missing_column),
+  expect_error(get_qk_coord(df_missing_column),
                regexp = paste("Please ensure that the dataset contains",
                               "a column named 'quadkey'."))
   
@@ -21,21 +21,21 @@ test_that("Check existence of 'quadkey' column", {
   df_with_column <- data.frame(quadkey = c("abc", "def", "ghi"))
   
   # Test case where 'quadkey' column is present
-  expect_silent(extract_qk_coord(grid$data))
+  expect_silent(get_qk_coord(grid$data))
 })
 
-# Tile XY coordinates - extract_tile_coord()
+# Tile XY coordinates - get_tile_coord()
 test_that("Check existence of 'tileX' and 'tileY' columns", {
   # Create a dataframe without 'tileX' and 'tileY' columns
   df_missing_columns <- data.frame(other_column = c(1, 2, 3))
   
   # Test case where 'tileX' and 'tileY' columns are missing
-  expect_error(extract_tile_coord(df_missing_columns, level = 6),
+  expect_error(get_tile_coord(df_missing_columns, level = 6),
                regexp = paste("Please ensure that the dataset contains",
                               "columns named 'tileX' and 'tileY'"))
   
   # Test case where 'tileX' and 'tileY' columns are present
-  expect_silent(extract_tile_coord(grid$data, level = 12))
+  expect_silent(get_tile_coord(grid$data, level = 12))
 })
 
 # Both
@@ -47,10 +47,10 @@ test_that("Check output dimensions", {
   
   # +1 is because I started counting by 0 the first element
   
-  expect_equal(nrow(extract_qk_coord(grid$data)),
+  expect_equal(nrow(get_qk_coord(grid$data)),
                ((grid$num_rows+1) * (grid$num_cols+1)))
   
-  expect_equal(nrow(extract_tile_coord(grid$data, level = 12)),
+  expect_equal(nrow(get_tile_coord(grid$data, level = 12)),
                ((grid$num_rows+1) * (grid$num_cols+1)))
 })
 
