+++ R/apply_weekly_lag.R
@@ -22,7 +22,7 @@
 #'
 #' @examples
 #'
-#' data <- result_read_all_files[1:500,]
+#' data <- result_read_fb_mobility_data[1:500,]
 #' apply_weekly_lag(data)
 #'
 apply_weekly_lag <- function(data){
