+++ man/get_qk_coord.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/extract_coords.R
-\name{extract_qk_coord}
-\alias{extract_qk_coord}
-\title{Extract lat/long coordinates from the QuadKey number}
+% Please edit documentation in R/get_coords.R
+\name{get_qk_coord}
+\alias{get_qk_coord}
+\title{Get lat/long coordinates from the QuadKey number}
 \usage{
-extract_qk_coord(data)
+get_qk_coord(data)
 }
 \arguments{
 \item{data}{A dataframe with a column named quadkey}
@@ -27,7 +27,7 @@ grid <- create_qk_grid(xmin = -59,
                        level = 6)
                        
 # quadkey column in grid$data converted to geographic coordinates
-grid_coords <- extract_qk_coord(data = grid$data)
+grid_coords <- get_qk_coord(data = grid$data)
 
 plot(grid_coords)
 }
