### 2024-01-28 75a1c76
Change create_raster() to create_stars_raster() (ropensci/software-review#619)
