+++ man/create_stars_raster.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/create_raster.R
-\name{create_raster}
-\alias{create_raster}
+% Please edit documentation in R/create_stars_raster.R
+\name{create_stars_raster}
+\alias{create_stars_raster}
 \title{Create a stars raster}
 \usage{
-create_raster(template, nx, ny, data, var)
+create_stars_raster(template, nx, ny, data, var)
 }
 \arguments{
 \item{template}{A spatial dataset (sf) with the polygon grid used as template}
@@ -44,11 +44,11 @@ raster, even in the presence of NAs.
                    dplyr::inner_join(data_provided,
                     by = c('quadkey' ))
 
- raster <-  create_raster(template = data_raster,
-                         nx = grid$num_cols + 1,
-                         ny = grid$num_rows + 1,
-                         data = data_raster,
-                         var = 'variable')
+ raster <-  create_stars_raster(template = data_raster,
+                                nx = grid$num_cols + 1,
+                                ny = grid$num_rows + 1,
+                                data = data_raster,
+                                var = 'variable')
 
 
 }
