### 2024-01-30 a6bdff9
Add warning() to regular_qk_grid() and make it return the dataset as the default behaviour (ropensci/software-review#619)
