### 2024-01-30 f105177
Fix style with styler::style_pkg() (ropensci/software-review#619)
