+++ tests/testthat/test-qk_to_latlong_functions.R
@@ -1,31 +1,39 @@
 # test quadkey to tile XY coordinates
 
 test_that("quadkey_to_tileXY is giving correct results", {
-  expect_equal(quadkey_to_tileXY('213')$tileX, 3)
-  expect_equal(quadkey_to_tileXY('213')$tileY, 5)
-  expect_equal(quadkey_to_tileXY('213')$level, 3)
+  expect_equal(quadkey_to_tileXY("213")$tileX, 3)
+  expect_equal(quadkey_to_tileXY("213")$tileY, 5)
+  expect_equal(quadkey_to_tileXY("213")$level, 3)
 })
 
 
 # test tile XY coordinates to pixel XY coordinates
 
 test_that("tileXY_to_pixelXY is giving correct results", {
-  expect_equal(tileXY_to_pixelXY(tileX = 3,
-                             tileY = 5)$pixelX, 768)
-  expect_equal(tileXY_to_pixelXY(tileX = 3,
-                             tileY = 5)$pixelY, 1280)
+  expect_equal(tileXY_to_pixelXY(
+    tileX = 3,
+    tileY = 5
+  )$pixelX, 768)
+  expect_equal(tileXY_to_pixelXY(
+    tileX = 3,
+    tileY = 5
+  )$pixelY, 1280)
 })
 
 
 # test pixels XY coordinates to map coordinates WG84 in degrees
 
 test_that("pixelXY_to_latlong is giving correct results", {
-  expect_equal(round(pixelXY_to_latlong(pixelX = 768,
-                                pixelY = 1280,
-                                level = 3)$lat, digits = 2), -40.98)
-  expect_equal(round(pixelXY_to_latlong(pixelX = 768,
-                                pixelY = 1280,
-                                level = 3)$lon, digits = 2), -45)
+  expect_equal(round(pixelXY_to_latlong(
+    pixelX = 768,
+    pixelY = 1280,
+    level = 3
+  )$lat, digits = 2), -40.98)
+  expect_equal(round(pixelXY_to_latlong(
+    pixelX = 768,
+    pixelY = 1280,
+    level = 3
+  )$lon, digits = 2), -45)
   # Test with valid level and pixel values
   result_valid <- pixelXY_to_latlong(100, 150, 5)
   expect_is(result_valid, "list")
@@ -35,59 +43,83 @@ test_that("pixelXY_to_latlong is giving correct results", {
 # Check that errors appear in the correct cases - pixelXY_to_latlong
 
 test_that("Check level detail - Negative level", {
-  expect_error(pixelXY_to_latlong(pixelX = 768,
-                                  pixelY = 1280,
-                                  level = -3),
-               regexp = paste("The level of detail should be",
-                              "an integer between 1 and 23"))
+  expect_error(pixelXY_to_latlong(
+    pixelX = 768,
+    pixelY = 1280,
+    level = -3
+  ),
+  regexp = paste(
+    "The level of detail should be",
+    "an integer between 1 and 23"
+  )
+  )
 })
 
 test_that("Check level detail - Greater than 23", {
-  expect_error(pixelXY_to_latlong(pixelX = 768,
-                                  pixelY = 1280,
-                                  level = 44),
-               regexp = paste("The level of detail should be",
-                              "an integer between 1 and 23"))
+  expect_error(pixelXY_to_latlong(
+    pixelX = 768,
+    pixelY = 1280,
+    level = 44
+  ),
+  regexp = paste(
+    "The level of detail should be",
+    "an integer between 1 and 23"
+  )
+  )
 })
 
 test_that("Check level detail - Use of decimals", {
-  expect_error(pixelXY_to_latlong(pixelX = 768,
-                                  pixelY = 1280,
-                                  level = 4.7),
[…]
