### 2024-02-01 c27e452
Make vignette C reproducible, update polygon_to_raster,  create_stars_raster and their tests (ropensci/software-review#619)
