### 2024-02-01 95100ee
Delete old image from vignette C (ropensci/software-review#619)
