### 2024-02-01 8c7944a
Add function quadkey_to_polygon and tests (ropensci/software-review#619)
