### 2024-02-03 4bd5cc9
Reintroduce n_crisis as fake variable to make apply_weely_lag example work. Minor fix test (ropensci/software-review#619)
