+++ vignettes/C-from_quadkey_identified_facebook_mobility_data_to_rasters.Rmd
@@ -51,12 +51,13 @@ files <- read_fb_mobility_files(path_to_csvs = paste0(system.file("extdata",
                                                            package = "quadkeyr"), "/"),
                                 colnames = c("lat", "lon", 
                                              "quadkey", "date_time", 
-                                             "percent_change"),
+                                             "n_crisis", "percent_change"),
                                 coltypes = list(
                                   lat = 'd',
                                   lon = 'd',
                                   quadkey = 'c',
                                   date_time = 'T',
+                                  n_crisis = 'c',
                                   percent_change = 'c')) 
 
 files
