+++ NAMESPACE
@@ -19,6 +19,7 @@ export(pixelXY_to_tileXY)
 export(polygon_to_raster)
 export(qkmap_app)
 export(quadkey_to_latlong)
+export(quadkey_to_polygon)
 export(quadkey_to_tileXY)
 export(read_fb_mobility_files)
 export(regular_qk_grid)
