### 2024-02-05 1f4e429
Change 'level of detail' argument to 'zoom' in multiple functions + other minor changes in the examples (ropensci/software-review#619)
