+++ tests/testthat/test-grid_to_polygon.R
@@ -5,7 +5,7 @@ grid <- create_qk_grid(
   xmax = -57,
   ymin = -35,
   ymax = -34,
-  level = 11
+  zoom = 11
 )
 
 grid_coords <- get_qk_coord(data = grid$data)
