+++ R/apply_weekly_lag.R
@@ -27,7 +27,7 @@
 #'                                 colnames = c("lat", "lon", 
 #'                                              "quadkey", "date_time", 
 #'                                              "n_crisis", "percent_change"),
-#'                                  coltypes = list(
+#'                                 coltypes = list(
 #'                                                 lat = 'd',
 #'                                                 lon = 'd',
 #'                                                 quadkey = 'c',
