+++ R/map_functions.R
@@ -9,23 +9,24 @@
 #' \href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
 #' Microsoft Bing Maps Tile System documentation}
 #'
-#' @param level Level of detail, from 1 (lowest detail) to 23 (highest detail).
+#' @param zoom Zoom or level of detail, 
+#' from 1 (lowest detail) to 23 (highest detail).
 #'
 #' @return The map width and height in pixels.
 #' @export
 #'
 #' @examples
 #'
-#' mapsize(level = 6)
+#' mapsize(zoom = 6)
 #'
-# At the lowest level of detail (level 1), the map is 512 x 512 pixels.
-# At each successive level of detail, the map width and height grow by a
+# At the lowest zoom level (level 1), the map is 512 x 512 pixels.
+# At each successive zoom level, the map width and height grow by a
 # factor of 2.
 # The pixel at the lower right corner of the map has pixel coordinates
-# (width-1, height-1) or ((256*2^level)–1, (256 * 2^level)–1).
+# (width-1, height-1) or ((256*2^zoom)–1, (256 * 2^zoom)–1).
 # Please, read the official documentation for more details.
-mapsize <- function(level) {
-  return(256 * 2^level)
+mapsize <- function(zoom) {
+  return(256 * 2^zoom)
 }
 
 
@@ -56,10 +57,10 @@ clip <- function(n, min_value, max_value) {
   return(min(max(n, min_value), max_value))
 }
 
-#' Ground resolution at a specified latitude and level of detail
+#' Ground resolution at a specified latitude and zoom level
 #'
 #' @description Determines the ground resolution (in meters per pixel) at a
-#' specified latitude and level of detail.
+#' specified latitude and zoom level.
 #' For further information, refer to the Microsoft Bing Maps Tile System
 #' documentation.
 #'
@@ -67,8 +68,10 @@ clip <- function(n, min_value, max_value) {
 #' \href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
 #' Microsoft Bing Maps Tile System documentation}
 #'
-#' @param latitude Latitude (in degrees) at which to measure the ground resolution.
-#' @param level Level of detail, from 1 (lowest detail) to 23 (highest detail).
+#' @param latitude Latitude (in degrees) at which
+#'  to measure the ground resolution.
+#' @param zoom Zoom or level of detail, 
+#' from 1 (lowest detail) to 23 (highest detail).
 #'
 #' @return the ground resolution (meters / pixel)
 #' @export
@@ -77,10 +80,10 @@ clip <- function(n, min_value, max_value) {
 #'
 #' ground_res(
 #'   latitude = 0,
-#'   level = 6
+#'   zoom = 6
 #' )
 #'
-ground_res <- function(latitude, level) {
+ground_res <- function(latitude, zoom) {
 
   # Values in Microsoft Bing Tile System Documentation
   min_latitude <- -85.05112878
@@ -90,7 +93,7 @@ ground_res <- function(latitude, level) {
   earth_radius <- 6378137
 
   latitude <- clip(latitude, min_latitude, max_latitude)
-  mapsize <- mapsize(level)
+  mapsize <- mapsize(zoom)
 
   return(cos(latitude * pi / 180) * 2 * pi * earth_radius / mapsize)
 }
@@ -107,7 +110,8 @@ ground_res <- function(latitude, level) {
 #' Microsoft Bing Maps Tile System documentation}
 #'
 #' @param latitude Latitude (in degrees) at which to measure the map scale.
[…]
