+++ R/data.R
@@ -5,7 +5,7 @@
 #' @format ## `data_provided`
 #' A data frame with 360 rows and 2 columns:
 #' \describe{
-#'   \item{quadkey}{QuadKey number as character}
+#'   \item{quadkey}{QuadKey as character}
 #'   \item{variable}{Numeric variable}
 #' }
 "data_provided"
@@ -17,9 +17,9 @@
 #' @format ## `result_read_fb_mobility_data`
 #' A data frame with 134,492 rows and 9 columns:
 #' \describe{
-#'   \item{lat}{Latitude of QuadKey centroid}
-#'   \item{lon}{Longitude of QuadKey centroid}
-#'   \item{quadkey}{QuadKey number as character}
+#'   \item{lat}{Latitude of the QuadKey centroid}
+#'   \item{lon}{Longitude of the QuadKey centroid}
+#'   \item{quadkey}{QuadKey as character}
 #'   \item{country}{Country name}
 #'   \item{date_time}{Date in format %Y-%m-%d %H%M}
 #'   \item{n_crisis}{Variable}

[…]
