### 2024-02-05 f0ba3ad
Update app to include the QuadKey polygon (ropensci/software-review#619)
