+++ inst/app/app.R
@@ -12,7 +12,7 @@ qkmap_app <- function(...){
                      bslib::layout_sidebar(class = 'p-0',
                                            sidebar = bslib::sidebar(width = 300,
                                                                     shiny::textInput('qk',
-                                                                                     'Visualize the upper-left coordinate of the QuadKey',
+                                                                                     'Visualize a specific QuadKey',
                                                                                      placeholder = "complete and click search"),
                                                                     shiny::actionButton('search', 'Search'),
                                                                     shiny::hr(),
@@ -22,8 +22,8 @@ qkmap_app <- function(...){
     bslib::nav_panel(title = 'Grid',
                      bslib::layout_sidebar(class = 'p-0',
                                            sidebar = bslib::sidebar(
-                                             shiny::p("Write the coordinates in decimal degrees"),
-                                             
+                                             shiny::p(paste("Write the bounding box",
+                                                            "coordinates in decimal degrees")),
                                              bslib::layout_columns(
                                                shiny::textInput('xmin', 'xmin'),
                                                shiny::textInput('ymin', 'ymin')
@@ -53,13 +53,13 @@ qkmap_app <- function(...){
   server <- function(input, output, session){
     
     # TAB 1: QuadKey
+    # Get the QuadKey Coorinates
     qk_results <-  shiny::eventReactive(input$search,
                                         ignoreNULL = FALSE,{
                                           shiny::validate(
                                             shiny::need(input$qk,
                                                         "Please, write a QuadKey number and press 'Search'")
                                           )
-                                          
                                           tile <- quadkeyr::quadkey_to_tileXY(as.character(input$qk))
                                           
                                           pixel <- quadkeyr::tileXY_to_pixelXY(tileX = tile$tileX,
@@ -77,7 +77,7 @@ qkmap_app <- function(...){
                                                       lat = coords$lat,
                                                       lon = coords$lon))
                                         })
-    
+    # Data table with QuadKey Info
     output$qkvalues <- DT::renderDT({
       
       dataqk <- t(data.frame(c('TileX',
@@ -104,20 +104,40 @@ qkmap_app <- function(...){
       
     })
     
+    qk_polygon <- shiny::eventReactive(input$search,
+                                       ignoreNULL = FALSE,{
+                        shiny::validate(
+        shiny::need(input$qk,
+                    "Please, write a QuadKey number and press 'Search'")
+      )
+                       quadkey_to_polygon(as.character(input$qk))                  
+                       })
+    
     # Quadkey location map
     output$mapqk <- leaflet::renderLeaflet({
-      
+      # get polygon bounding box
+      bbx <-  sf::st_bbox(qk_polygon())
+      # Map tab 1, QuadKey polygon and upper-left corner coordinate
       leaflet::leaflet() |>
         leaflet::addTiles() |>
+        leaflet::addMeasure(primaryLengthUnit="kilometers",
+                            secondaryLengthUnit="kilometers") |> 
+        leaflet::addPolygons(data = qk_polygon(),
+                             group = 'polygon',
+                             layerId = ~quadkey, # https://rstudio.github.io/leaflet/showhide.html
+                             color = "red",
+                             fillColor = NA,
+                             label =  ~as.character(input$qk)) |> 
         leaflet::addMarkers(data = qk_results(),
                             lat = ~lat,
                             lng = ~lon,
-                            popup = ~paste(qk_results()$lat, 
+                             ~paste(qk_results()$lat, 
[…]
