+++ vignettes/A-bing_map_tile_system_functions.Rmd
@@ -171,7 +171,7 @@ quadkey_to_latlong("2103")
 ```
 
 
-### 3. Understanding the Process of Converting Coordinates
+### 3. Caveats when converting coordinates
 
 Given the process of converting geographic coordinates (latitude/longitude)
 to QuadKeys, one might expect that the conversion back to
