### 2024-02-05 63c8932
Remove don't run from format_fb_data() example (ropensci/software-review#619)
