### 2024-02-05 #619 @flor14
Update: I am moving forward with this. I am planning some changes in how the documentation is organized, so it is going to take me a bit longer than I expected originally.
