### 2024-02-07 7d5d3f9
Update first vignette including workflow, renaming .Rmd file, pkgdown.yaml article order and improvements in the technical ddescription of the functions (ropensci/software-review#619)
