+++ vignettes/quadkey_to_sf_conversion.Rmd
@@ -1,8 +1,10 @@
 ---
-title: "Bing Maps Tile System functions"
-output: rmarkdown::html_vignette
+title: "From QuadKey to Simple Features data.frame and other conversions"
+output:
+  html_document:
+    number_sections: true
 vignette: >
-  %\VignetteIndexEntry{Bing Maps Tile System functions}
+  %\VignetteIndexEntry{From QuadKey to Simple Features data.frame and other conversions}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -16,6 +18,7 @@ knitr::opts_chunk$set(
 
 ```{r setup, echo = FALSE, message=FALSE, warning=FALSE}
 library(quadkeyr)
+library(ggplot2)
 ```
 
 > Please, visit the [README](https://fernandez-lab-wsu.github.io/quadkeyr/) 
@@ -24,9 +27,58 @@ library(quadkeyr)
 Most of the function and arguments names conserved from the documentation.
 `quadkeyr` functions are the R version of the ones described in [Bing Maps Tile System webpage](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
 
-Tile maps are composed of pixels that are grouped into tiles. Later, the tiles
-are converted to QuadKeys to optimize map performance, among other benefits 
-described in detail in the documentation.
+
+## Basic workflow
+
+```{r w1, echo = FALSE, out.width= "90%", fig.align='center'}
+knitr::include_graphics("../man/figures/workflow_quadkey.png")
+```
+
+You can convert any QuadKey to
+a [Simple Features object](https://r-spatial.github.io/sf/articles/sf1.html#sf-objects-with-simple-features)
+`sf POINT data.frame` or `sf POLYGON data.frame` 
+using the functions `quadkey_to_latlong` and `quadkey_to_polygon`. 
+It's important to notice that the geographic coordinates of the point 
+will always correspond to the upper-left corner of the polygon
+for the same QuadKey.
+
+```{r}
+qtll <- quadkey_to_latlong("2103")
+qtll
+```
+
+```{r}
+qtp <- quadkey_to_polygon("2103")
+qtp
+```
+
+```{r qk, out.width="70%", fig.align='center'}
+ggplot() +
+  geom_sf(data = qtp, color = 'red', fill = NA, linewidth = 2.5) +
+  geom_sf(data = qtll, color = 'blue', size = 4) + 
+  ggtitle(paste("QuadKey", qtll$quadkey)) +
+  theme_minimal()
+```
+
+## Advanced use
+
+The conversion from QuadKey to geographic coordinates
+involves a series of smaller transformations that
+relate to the structure of Tile Maps. 
+All of these intermediary functions, 
+as well as those facilitating the reverse conversion,
+are available for use within the `quadkeyr` package.
+
+### Tile Maps: QuadKeys, tiles, pixels and geographic coordinates
+
+> Note: In the official documentation, 
+> the term 'level of detail' is equivalent to 
+> what is here referred to as 'zoom level'.
+
+Tile maps are composed of pixels that are grouped into tiles. 
+Later, the tiles are converted to QuadKeys to optimize map performance,
+among other benefits described in detail in 
+[the documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
 
 Pixels and tiles are expressed as two-dimensional coordinates - (`pixelX`, 
 `pixelY`) and (`tileX`, `tileY`) - but Quadkeys are one-dimensional
@@ -45,8 +97,9 @@ knitr::include_graphics('bing_pixel.jpg')
 ```
 
 
[…]
