+++ README.Rmd
@@ -42,7 +42,7 @@ spatial position as a sequence of characters. They provide an efficient method
 to address and retrieve specific map tiles, facilitating rapid display within 
 mapping applications such as [Microsoft Bing Maps](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system).
 
-```{r qk, echo = FALSE, out.width= "90%", fig.align='center', fig.cap= "The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage."}
+```{r qk, echo = FALSE, out.width= "70%", fig.align='center', fig.cap= "The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage."}
 knitr::include_graphics("man/figures/quadkeys.jpg")
 ```
 
@@ -51,9 +51,17 @@ knitr::include_graphics("man/figures/quadkeys.jpg")
 
 The goal of `quadkeyr` is to:  
   
-1. [**Provide Bing Maps Tile System functions**](https://fernandez-lab-wsu.github.io/quadkeyr/articles/quadkey_conversion.html)
-R functions to convert geographical coordinates 
-to and from QuadKeys as described in the [official documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system)  
+1. [**Convert a QuadKey to a simple features data.frame (and more)**](https://fernandez-lab-wsu.github.io/quadkeyr/articles/quadkey_conversion.html)
+`quadkeyr` provides functions to convert a QuadKey to
+a `sf POINT data.frame` or `sf POLYGON data.frame`.
+Additionally, all the R functions described 
+in the [official documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system)  
+to convert QuadKeys to and from tiles, pixels and geographic coordinates 
+are available.
+
+```{r w1, echo = FALSE, out.width= "90%", fig.align='center'}
+knitr::include_graphics("man/figures/workflow_quadkey.png")
+```
   
 2. [**Create rasters based on QuadKeys**](https://fernandez-lab-wsu.github.io/quadkeyr/articles/get_grid_from_quadkeys.html)
 Complete a grid of QuadKeys within a specified area and level of detail, 
