### 2024-02-10 e5db815
Multiple updates regarding vignette 2: conversion dataframe to raster (ropensci/software-review#619) - Rename .Rmd vignette and add it in order to _pkgdown.yaml - First version of functions add_regular_polygon_grid and get_regular_polygon_grid with multiple functions updates to make them work - Add basic workflow with image in vignette
