+++ R/regular_grid.R
@@ -20,6 +20,8 @@
 #' regular_qk_grid(qtll)
 #'
 regular_qk_grid <- function(data) {
+  # I need the data.frame to be class simple features to can 
+  # estimate the bounding box
   bbox <- sf::st_bbox(data) 
 
   # I assume that the all the QuadKeys correspond to the same zoom level.
@@ -42,18 +44,21 @@ regular_qk_grid <- function(data) {
 
   # Select the QuadKeys that are missing in the original grid
   qk_missing <- grid$data |>
-    dplyr::anti_join(data, by = "quadkey")
+    dplyr::anti_join(data, by = "quadkey") |> 
+    dplyr::select(-tileX, -tileY)
   
   # Convert the QuadKeys to coordinates
   grid_coords <- quadkey_to_latlong(qk_missing$quadkey)
 
+  
+  if(ncol(grid_coords) == ncol(data)){
   # Add the missing QuadKey coordinates to the original dataset
   data <- rbind(grid_coords, data)
-
-  # I will need the tiles for grid_to_polygon
-  for (i in seq_len(nrow(data))) {
-    data[i, "tileX"] <- quadkey_to_tileXY(data$quadkey[i])$tileX
-    data[i, "tileY"] <- quadkey_to_tileXY(data$quadkey[i])$tileY
+  }else{
+    # this function will introduce NAs if there are columns 
+    # different than quadkey and geometry
+    # probably used inside add_regular_polygon_grid()
+  data <- dplyr::bind_rows(grid_coords, data)
   }
 
   return(list(
@@ -62,3 +67,80 @@ regular_qk_grid <- function(data) {
     num_cols = grid$num_cols
   ))
 }
+
+
+
+#' Complete a data.frame with the quadkeys needed to have a regular grid.
+#'
+#' @description
+#' This function estimates the bounding box of the quadkeys given in the 
+#' quadkey column and adds rows to complete the quadkeys and the geometry 
+#' needed to create a regular grid.
+#' All other columns for the introduced quadkeys will be filled with NAs.
+#' 
+#' @param data A data.frame with a quadkey column.
+#' 
+#' @return An sf POLYGON data.frame with all the quadkeys within
+#' the bounding box of the provided quadkeys.
+#' The columns quadkey and geometry are returned for all the grid,
+#' and the values of the introduced quadkeys will be NA
+#' for the rest pf the variables.
+#'  
+#' @export
+#'
+#' @examples
+add_regular_polygon_grid <- function(data){
+  
+  
+  # # read the file with the data
+  # path <- paste0(system.file("extdata", package = 'quadkeyr'), "/cityA_2020_04_15_0000.csv")
+  # data <- read.csv(path)
+  # data <- format_fb_data(data)
+  # 
+   sf_grid <- get_qk_coord(data)
+   reggrid <- regular_qk_grid(sf_grid)
+   add_polygrid <- grid_to_polygon(reggrid$data)
+   return(list(data = add_polygrid,
+          num_cols = reggrid$num_cols,
+          num_rows = reggrid$num_rows))
+
+}
+  
+
+#' Create a data.frame with all the quadkeys needed to
+#' generate a regular grid.
+#' 
+#' @description
+#' This function estimates the bounding box of the quadkeys given in the 
+#' quadkey column and adds rows to complete the quadkeys and the geometry 
+#' needed to create a regular grid.
+#'
+#' @param data A data.frame with a quadkey column.
+#'
+#' @return An sf POLYGON data.frame with all the quadkeys within
+#' the bounding box of the provided quadkeys.
+#' Only the columns quadkey and geometry are returned.
+#' 
+#' @export
+#'
+#' @examples
+#' 
[…]
