+++ R/grid_to_polygon.R
@@ -1,3 +1,121 @@
+#' Convert a grid of QuadKeys to square polygons
+#'
+#' @description The main argument of this function, the grid of
+#' geographic coordinates (lat/long WG84) represents the upper-left
+#' corner of the QuadKey.
+#' To transform these coordinates into square polygons, the function
+#' supplements the grid by adding a row and column of tiles. These points
+#' introduce QuadKeys located at the border of the area
+#' (complete_grid_for_polygons).
+#' The function constructs the polygons using all the points of the grid.
+#' Note that it's possible to associate each QuadKey with its square polygon.
+#'
+#' @param data A spatial dataset (sf) with a quadkey and POINT geometry column.
+#' If the columns tileX and tileY are included in the input dataset,
+#' they will be used by complete_grid_for_polygons().
+#' If not, the tile coordinates will be calculated internally.
+#'
+#' @seealso \code{\link{complete_grid_for_polygons}}
+#'
+#' @return A sf POLYGON data.frame with a quadkey column.
+#'
+#' @export
+#'
+#' @examples
+#' grid <- create_qk_grid(
+#'   xmin = -59,
+#'   xmax = -57,
+#'   ymin = -35,
+#'   ymax = -34,
+#'   zoom = 11
+#' )
+#'
+#' grid_coords <- get_qk_coord(data = grid$data)
+#'
+#' polygrid <- grid_to_polygon(grid_coords)
+#' polygrid
+grid_to_polygon <- function(data) {
+  if (!("sf" %in% class(data))) {
+    stop("The dataset should be of class 'sf'")
+  }
+  
+  # Convert the QuadKeys to tile coordinates
+  # if that columns aren't present in the data
+  if (!("tileX" %in% colnames(data) | "tileY" %in% colnames(data))) {
+    message(paste(
+      "The 'tileX' and 'tileY' columns have been generated",
+      "using the 'quadkey_to_tileXY' function."
+    ))
+    
+    # In the case that one of this columns is not present, 
+    # calculate both again.
+    data$tileX <- NA
+    data$tileY <- NA
+    
+    data <- data |>
+      dplyr::rowwise() |>
+      dplyr::mutate(
+        tileX = quadkey_to_tileXY(.data$quadkey)$tileX,
+        tileY = quadkey_to_tileXY(.data$quadkey)$tileY
+      )
+  }
+  
+  extragrid <- complete_grid_for_polygons(data)
+  
+  extragrid <- get_tile_coord(extragrid,
+                              zoom <- unique(nchar(data$quadkey))
+  )
+  
+  if(ncol(data) == ncol(extragrid)){
+    # combines the new data with the extended grid of points
+    data <- rbind(data, extragrid)
+  }else{
+    # In the case that there are more columns in data, 
+    # keep the rows but adding NAs in the extragrid cells.
+    data <- dplyr::bind_rows(data, extragrid)
+  }
+  
+  polydata <- c() # https://github.com/r-spatial/sf/issues/354
+  
+  # The QuadKeys of interest are the ones that are not NA:
+  # the original QuadKeys of the grid
+  subdata <- subset(data, !is.na(data$quadkey))
+  
+  for (i in seq_len(nrow(subdata))) {
+    x <- subdata[i, ]$tileX
+    y <- subdata[i, ]$tileY
+    
+    # This point will always be a QuadKey in the dataframe
+    a <- data[data$tileX == x & data$tileY == y, ]
+    
+    # b, c and d can be part of the extended grid
+    b <- data[data$tileX == x & data$tileY == (y + 1), ]
+    
[…]
