+++ _pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   navbar: ~
   contents:
   - quadkey_to_sf_conversion
-  - B-from_quadkey_identified_data_to_raster
+  - quadkey_identified_data_to_raster
   - C-from_quadkey_identified_facebook_mobility_data_to_rasters
   - D-quadkey_visualization_app
 

[…]
