### 2024-02-10 9eac5fb
Add test function polygon_to_raster (ropensci/software-review#619)
