### 2024-02-11 b7fea38
Update vignettes 2 and 3, readme, tests and format (ropensci/software-review#619) - pkgdown articles where corrected by name and well organized - Fix tests after changes in the basic workflow - Add updated images - Update readme - Text improvements
