### 2024-02-13 f4ecb2d
close #2 as part of ropensci/software-review#619
