+++ tests/testthat/test-qk_to_latlong_functions.R
@@ -1,18 +1,41 @@
-# test quadkey to tile XY coordinates
+# Test QuadKey to tile XY coordinates conversion
 
 test_that("quadkey_to_tileXY is giving correct results", {
   expect_equal(quadkey_to_tileXY("213")$tileX, 3)
   expect_equal(quadkey_to_tileXY("213")$tileY, 5)
   expect_equal(quadkey_to_tileXY("213")$zoom, 3)
-  
   expect_equal(quadkey_to_tileXY('0313102310')$tileX, 486)
   expect_equal(quadkey_to_tileXY('0313102310')$tileY, 332)
   expect_equal(quadkey_to_tileXY('0313102310')$zoom, 10)
 })
 
+test_that("Error: QuadKeys should be strings", {
+  # Test with a character vector
+  expect_silent(quadkey_to_tileXY("322"))
+  # Test with a number
+  expect_error(quadkey_to_tileXY(322), 
+               "Please provide a QuadKey as a single string")
+  # Test with a vector
+  expect_error(quadkey_to_tileXY(c("322", "123")), 
+               "Please provide a QuadKey as a single string")
+})
 
-# test tile XY coordinates to pixel XY coordinates
+test_that("Error: QuadKeys contain only the numbers '0', '1', '2', or '3'", {
+  # Test with valid input
+  valid_quadkey <- "0123"
+  expect_silent(quadkey_to_tileXY(valid_quadkey))
+  
+  # Test with invalid input character
+  invalid_quadkey <- "123f"
+  expect_error(quadkey_to_tileXY(invalid_quadkey), 
+               "QuadKeys can contain only the numbers '0', '1', '2', or '3'")
+  # Test with invalid input number
+  invalid_quadkey2 <- "1234"
+  expect_error(quadkey_to_tileXY(invalid_quadkey2), 
+               "QuadKeys can contain only the numbers '0', '1', '2', or '3'")
+})
 
+# Test tile XY coordinates to pixel XY coordinates
 test_that("tileXY_to_pixelXY is giving correct results", {
   expect_equal(tileXY_to_pixelXY(
     tileX = 3,
@@ -102,32 +125,56 @@ test_that("pixelXY_to_latlong handles invalid pixelX and pixelY values", {
 test_that("quadkey_to_latlong function works as expected", {
   quadkeys <- c("213", "312", "213")
 
-  # Test 1: Check for duplicated quadkeys
+  # Check for duplicated quadkeys
   expect_error(
     quadkey_to_latlong(quadkeys),
     "Please, remove duplicated QuadKeys"
   )
 
-  # Test 2: Check for quadkeys with different digit lengths
+  # Check for quadkeys with different digit lengths
   expect_error(
     quadkey_to_latlong(c("2333", "123", "3310021")),
     "All the QuadKeys should have the same number of digits"
   )
 
-  # Test 3: Check if the returned object is an sf object
+  # Check if the returned object is an sf object
   result_sf <- quadkey_to_latlong(c("212", "213"))
   expect_true("sf" %in% class(result_sf),
     info = "Output should be of class 'sf'"
   )
 
-  # Test 4: Check if sf object contains expected columns
+  # Check if sf object contains expected columns
   expect_equal(colnames(result_sf), c("quadkey", "geometry"),
     info = "SF object should have 'quadkey' and 'geometry' columns"
   )
 
-  # Test 5: Check that quadkey_to_latlong() is giving correct values
+  # Check that quadkey_to_latlong() is giving correct values
   expect_equal(sf::st_coordinates(quadkey_to_latlong("0313102310"))[1], 
                -9.140625, tolerance = 1e-06)
[…]
