### 2024-02-13 08898ad
Updates in *_to_polygon() functions (ropensci/software-review#619)  - Now both *_to_polygon functions have their own file in R/  and tests and where removed from their original files.  - Add comment in 2nd vignette about the use of quadkey_df_to_polygon().  - Update app images.
