+++ R/grid_to_polygon.R
@@ -193,49 +193,3 @@ complete_grid_for_polygons <- function(data) {
   return(extragrid)
 }
 
-#' Convert a QuadKey into a square polygon
-#'
-#' This functions creates a sf class polygon from a QuadKey.
-#'
-#' @param quadkey The QuadKey as a string
-#'
-#' @return A spatial dataset (sf) with a quadkey and POLYGON geometry column.
-#' @export
-#'
-#' @examples
-#' 
-#' quadkey_to_polygon('213')
-#' 
-quadkey_to_polygon <- function(quadkey){
-  
-  tileX <- quadkey_to_tileXY(quadkey)$tileX
-  tileY <- quadkey_to_tileXY(quadkey)$tileY
-  
-  # I add the other 3 points I need to complete the polygon
-  # using the upper left coordinates of the closer tiles.
-  polygon <- expand.grid(tileX = c(tileX, tileX+1),
-              tileY = c(tileY, tileY+1)) |> 
-    dplyr::rowwise() |> 
-    dplyr::mutate(pixelX = tileXY_to_pixelXY(.data$tileX, # conversion to coords
-                                      .data$tileY)$pixelX,
-           pixelY = tileXY_to_pixelXY(.data$tileX,
-                                      .data$tileY)$pixelY) |> 
-    dplyr::mutate(lat = pixelXY_to_latlong(.data$pixelX, 
-                                           .data$pixelY,
-                                           nchar(quadkey))$lat,
-           lon = pixelXY_to_latlong(.data$pixelX, 
-                                    .data$pixelY,
-                                    nchar(quadkey))$lon) |> 
-    sf::st_as_sf(coords = c("lon", "lat"), # class sf
-             crs = 4326)
-    
-  # Create polygon https://github.com/r-spatial/sf/issues/243
- quadkey_polygon <-  polygon |>
-    sf::st_bbox() |>
-    sf::st_as_sfc() |> 
-    sf::st_sf() |> 
-    sf::st_set_geometry("geometry") |> 
-    cbind(quadkey)
- 
-  return(quadkey_polygon)
-}
