+++ tests/testthat/test-grid_to_polygon.R
@@ -24,8 +24,3 @@ test_that("Dataset is of class 'sf'", {
   )
 })
 
-test_that("quadkey_to_polygon gives the correct output", {
-  polygon <- quadkey_to_polygon('213')
-  expect_s3_class(polygon, "sf")
-  expect_equal(polygon$quadkey, "213")
-})

[…]
