### 2024-02-13 a7a5f32
Update vignettes, tests and fix +1 unnecesary use in create_stars_raster() (ropensci/software-review#619) - Improve and homogenize tests format. - Adds +1 as part of create_qk_grid() and it is removed in vignettes. - Update second vignette with a most representative case of use.
