+++ tests/testthat/test-regular_grid.R
@@ -2,19 +2,19 @@
 data <- quadkey_to_latlong(quadkey = c("213", "312", "123"))
 result <- regular_qk_grid(data)
 
-# Test cases for regular_qk_grid function
+# Test cases for `regular_qk_grid()`
 test_that("regular_qk_grid creates the expected dataframe", {
-  # Test 1: Check the structure of the output
+  # Check the structure of the output
   expect_s3_class(result$data, "sf")
-  expect_equal(result$num_rows, 2)
-  expect_equal(result$num_cols, 3)
+  expect_equal(result$num_rows, 3)
+  expect_equal(result$num_cols, 4)
   
-  # Test 2: Check if output data frame has expected columns
+  # Check if output data frame has expected columns
   expected_cols <- c("quadkey", "geometry")
   expect_equal(colnames(result$data), expected_cols,
                info = "Output data frame should have expected columns")
   
-  # Test 3: Check if the warning appears
+  # Check if the warning appears
   expect_warning(
     regular_qk_grid(result$data),
     paste("The grid is already complete,",
@@ -22,8 +22,7 @@ test_that("regular_qk_grid creates the expected dataframe", {
   )
 })
 
-# get_regular_grid()
-
+# Test cases for `get_regular_grid()`
 files <- read_fb_mobility_files(
   path_to_csvs = paste0(system.file("extdata",
                                     package = "quadkeyr"), "/"),
@@ -47,7 +46,7 @@ files <- read_fb_mobility_files(
 
 regular_grid <- get_regular_polygon_grid(data = files)
 
-test_that("get_regular_polygon_grid returns the expected output structure",
+test_that("`get_regular_polygon_grid()` returns the expected output structure",
           {
             # Check the structure of the output
             # Check if the result is a list
@@ -65,24 +64,25 @@ test_that("get_regular_polygon_grid returns the expected output structure",
             expect_type(regular_grid$num_rows, "double")
           })
 
-test_that("get_regular_polygon_grid returns the expected # of columns and rows",
+test_that("`get_regular_polygon_grid()` returns the expected # of columns and rows",
           {
             # Check if the number of cols and rows matches the expected values
-            expect_equal(regular_grid$num_cols, 17)
-            expect_equal(regular_grid$num_rows, 21)
+            expect_equal(regular_grid$num_cols, 18)
+            expect_equal(regular_grid$num_rows, 22)
           })
 
 
-test_that("get_regular_polygon_grid is providing a complete grid", {
+test_that("`get_regular_polygon_grid()` is providing a complete grid", {
   # Test 1: The QuadKeys are in relation to the grid dimensions
-  expect_equal((regular_grid$num_cols + 1) * (regular_grid$num_rows + 1),
+  expect_equal((regular_grid$num_cols) * (regular_grid$num_rows),
                length(regular_grid$data$quadkey)
   )
 })
 
+# Test cases for `add_regular_polygon_grid()`
 added_regular_grid <- add_regular_polygon_grid(data = files)
 
-test_that("add_regular_polygon_grid returns the expected output structure",
+test_that("`add_regular_polygon_grid()` returns the expected output structure",
           {
             # Check if the result is a list
             expect_type(added_regular_grid, "list")
@@ -99,14 +99,14 @@ test_that("add_regular_polygon_grid returns the expected output structure",
             expect_type(added_regular_grid$num_rows, "double")
           })
 
-test_that("add_regular_polygon_grid returns the expected # of col and rows",
+test_that("`add_regular_polygon_grid()` returns the expected # of col and rows",
           {
             # Check if the number of cols and rows matches the expected values
-            expect_equal(added_regular_grid$num_cols, 17)
[…]
