+++ tests/testthat/test-grid_to_polygon.R
@@ -1,5 +1,5 @@
-
-# example
+# Test cases for `grid_to_polygon()`
+# Mock data
 grid <- create_qk_grid(
   xmin = -59,
   xmax = -57,
@@ -18,9 +18,6 @@ polygrid
 # Check that the error is working when the data is not class 'sf'
 test_that("Dataset is of class 'sf'", {
   expect_silent(polygrid)
-
-  expect_error(grid_to_polygon(as.data.frame(grid_coords)),
-    regexp = "The dataset should be of class 'sf'"
-  )
+  expect_error(grid_to_polygon(as.data.frame(grid_coords)))
 })
 
