### 2024-02-13 157701d
General improvement of @examples and roxygen docs (ropensci/software-review#619)
