+++ R/get_coords.R
@@ -10,7 +10,7 @@
 #' @seealso \code{\link{pixelXY_to_latlong}}
 #'
 #' @return A `sf` POINT data.frame containing the tiles XY coordinates
-#' (tileX, tileY), the QuadKeys (quadkey), and a geometry column.
+#' (`tileX`, `tileY`), the QuadKeys (`quadkey`), and a `geometry` column.
 #'
 #' @export
 #'
@@ -61,15 +61,15 @@ get_qk_coord <- function(data) {
 #' @description Reads the tile XY coordinates and extracts the
 #' lat/long coordinates of the upper-left corner of the QuadKey.
 #'
-#' @param data A dataframe with columns named tileX and tileY
+#' @param data A dataframe with columns named `tileX` and `tileY`
 #' @param zoom Zoom or Level of detail, 
 #' from 1 (lowest detail) to 23 (highest detail).
 #'
 #' @seealso \code{\link{tileXY_to_pixelXY}}
 #' @seealso \code{\link{pixelXY_to_latlong}}
 #'
-#' @return  A spatial dataframe (sf) containing the tiles XY coordinates
-#' (tileX, tileY)  and a column for POINT geometry.
+#' @return  A `sf` POINT data.frame containing the tiles XY coordinates
+#' (`tileX`, `tileY`), a `quadkey` and a `geometry` column.
 #'
 #' @export
 #'
