+++ R/qk_to_latlong_functions.R
@@ -2,10 +2,20 @@
 #'
 #' @description  For further information, refer to the
 #' Microsoft Bing Maps Tile System documentation.
-#'
+#' 
+#' @details
+#' Converting latitude/longitude coordinates into a QuadKey 
+#' and then back to latitude/longitude won't yield identical values, 
+#' unless the initial latitude/longitude coordinates 
+#' correspond to the upper-left Quadkey's pixel and tile XY coordinates
+#' at the same zoom level. 
+#'  
+#' Understanding this distinction is crucial for 
+#' the accurate use of these functions in coordinate conversions.
+#'  
 #' For a detailed explanation on how to use this
 #' and other similar `quadkeyr` functions,
-#' see the vignette:
+#' read the the vignette:
 #' \url{https://fernandez-lab-wsu.github.io/quadkeyr/articles/
 #' quadkey_to_sf_conversion.html}
 #' 
@@ -67,9 +77,19 @@ return(list(
 #' For further information, refer to the Microsoft Bing Maps Tile System
 #' documentation.
 #' 
+#' @details
+#' Converting latitude/longitude coordinates into a QuadKey 
+#' and then back to latitude/longitude won't yield identical values, 
+#' unless the initial latitude/longitude coordinates 
+#' correspond to the upper-left Quadkey's pixel and tile XY coordinates
+#' at the same zoom level. 
+#'  
+#' Understanding this distinction is crucial for 
+#' the accurate use of these functions in coordinate conversions.
+#'  
 #' For a detailed explanation on how to use this
 #' and other similar `quadkeyr` functions,
-#' see the vignette:
+#' read the the vignette:
 #' \url{https://fernandez-lab-wsu.github.io/quadkeyr/articles/
 #' quadkey_to_sf_conversion.html}
 #'
@@ -113,9 +133,19 @@ tileXY_to_pixelXY <- function(tileX, tileY) {
 #' For further information, refer to the
 #' Microsoft Bing Maps Tile System documentation.
 #'
+#' @details
+#' Converting latitude/longitude coordinates into a QuadKey 
+#' and then back to latitude/longitude won't yield identical values, 
+#' unless the initial latitude/longitude coordinates 
+#' correspond to the upper-left Quadkey's pixel and tile XY coordinates
+#' at the same zoom level. 
+#'  
+#' Understanding this distinction is crucial for 
+#' the accurate use of these functions in coordinate conversions.
+#'  
 #' For a detailed explanation on how to use this
 #' and other similar `quadkeyr` functions,
-#' see the vignette:
+#' read the the vignette:
 #' \url{https://fernandez-lab-wsu.github.io/quadkeyr/articles/
 #' quadkey_to_sf_conversion.html}
 #'
@@ -182,13 +212,23 @@ pixelXY_to_latlong <- function(pixelX, pixelY, zoom) {
 #' For further information, refer to the Microsoft Bing Maps Tile System
 #' documentation.
 #' 
+#' @details
+#' Converting latitude/longitude coordinates into a QuadKey 
+#' and then back to latitude/longitude won't yield identical values, 
+#' unless the initial latitude/longitude coordinates 
+#' correspond to the upper-left Quadkey's pixel and tile XY coordinates
+#' at the same zoom level. 
+#'  
+#' Understanding this distinction is crucial for 
+#' the accurate use of these functions in coordinate conversions.
+#'  
[…]

[…]
