### 2024-02-13 f36d676
Fix toc vignettes and build pkgdown again - minor typos (ropensci/software-review#619)
