+++ R/data.R
@@ -1,25 +1,25 @@
-#' data_provided: fake dataset
+#' QuadKey-identified Dataset
 #'
-#' A data.frame simulating a variable asociated with QuadKeys.
+#' A data.frame simulating a QuadKey-identified dataset.
 #'
 #' @format ## `data_provided`
-#' A data frame with 360 rows and 2 columns:
+#' A data.frame with 360 rows and 2 columns:
 #' \describe{
-#'   \item{quadkey}{QuadKey as character}
+#'   \item{quadkey}{QuadKey as a string}
 #'   \item{variable}{Numeric variable}
 #' }
 "data_provided"
 
-#' `result_read_fb_mobility_data`: Fake dataset
+#' Dataset with (fake) Facebook mobility data
 #'
-#' A data frame similar to a the potential output of `read_fb_mobility_files`.
+#' A data.frame similar to a the potential output of `read_fb_mobility_files`.
 #'
 #' @format ## `result_read_fb_mobility_data`
 #' A data frame with 134,492 rows and 9 columns:
 #' \describe{
 #'   \item{lat}{Latitude of the QuadKey centroid}
 #'   \item{lon}{Longitude of the QuadKey centroid}
-#'   \item{quadkey}{QuadKey as character}
+#'   \item{quadkey}{QuadKey as a string}
 #'   \item{country}{Country name}
 #'   \item{date_time}{Date in format %Y-%m-%d %H%M}
 #'   \item{n_crisis}{Variable}
