+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: quadkeyr
-Title: Tools for converting QuadKeys used in Microsoft's Bing Maps Tile System into raster images
+Title: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images
 Version: 0.0.0.9000
 Authors@R: 
     person(given = "Florencia",
@@ -7,9 +7,9 @@ Authors@R:
            email = "florencia.dandrea@gmail.com", 
            role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-0041-097X"))
-Description: 'Quadkeyr' enables the generation of raster images based on QuadKeys, facilitating efficient integration of Bing Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
+Description: 'Quadkeyr'functins generate raster images based on QuadKey-identified data, facilitating efficient integration of Tile Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
 License: MIT + file LICENSE
-URL: <https://fernandez-lab-wsu.github.io/quadkeyr/>, <https://github.com/Fernandez-Lab-WSU/quadkeyr>
+URL: https://fernandez-lab-wsu.github.io/quadkeyr/, https://github.com/Fernandez-Lab-WSU/quadkeyr
 BugReports: https://github.com/Fernandez-Lab-WSU/quadkeyr/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
