### 2024-02-14 #619 @flor14
Hello,
Before addressing the last review, I want to make some general updates regarding the state of the package.
NOTE: When referencing changes made to items such as vignettes, that have received multiple changes in the text, I will link the last commit for easier identification.

## NEWS
### **PKGDOWN SITE (Technical changes)**
1. **Vignettes filenames**: There were some misunderstandings regarding the vignette names. The problem was that I was not including the articles in the `_pkgdown.yml`, so they appeared in a random order. Now, I have simplified and organized them as follows:
- `quadkey_to_sf_conversion`
- `quadkey_identified_data_to_raster`
- `facebook_mobility_csvs_to_raster_files`
- `quadkey_visualization_app`

I believe these names are much more representative of the content of the vignettes.
You can find the last version of the vignettes here: https://github.com/Fernandez-Lab-WSU/quadkeyr/tree/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes

2. **Number of sections:** @vincentvanhees, I copied and pasted the exactly same `YAML` you sent me, but for some reason, it was not working for me. After some reading, this worked:
```
---
title: "Generating a Raster Image from Quadkey-Identified Data"
output:
   html_document:
    toc : true
    number_sections: true
    toc_depth: 3
pkgdown:
  as_is: true
vignette: >
  %\VignetteIndexEntry{Generating a Raster Image from Quadkey-Identified Data}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---
```
You can find the last version of the vignettes here: https://github.com/Fernandez-Lab-WSU/quadkeyr/tree/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes

###  **PKGDOWN SITE (Changes in the articles)**
The major change was made in response to Maria Paula’s comment about adding a basic workflow. In particular, I think the major improvement is in the second vignette (`quadkey_identified_data_to_raster`):
- I have divided the three principal articles into a basic workflow section followed by the old documentation as an advanced use case.
- The basic workflow section contains a diagram that is also included in the README for each of the cases.
- There have been included more technical terms in all the documentation, for example, links to Simple Features (`sf`) and `stars` packages and the homogenization of the terminology. For example, to refer to a `sf` class data frame we call it `sf` POLYGON data.frame or `sf` POINT data.frame.
- All the examples of the vignettes are reproducible, as there have been added example files in `inst/extdata`.
Fernandez-Lab-WSU/quadkeyr@7d5d3f9c

### **NEW FUNCTIONS**
- Two new functions have been added that directly convert QuadKeys to polygons, `quadkey_to_polygon()` and `quadkey_df_to_polygon()`. One takes a string as an argument, while the other takes a data.frame. Both functions are included in a new file named `qk_to_polygon_functions.R`. Fernandez-Lab-WSU/quadkeyr@8c7944a8
- Two functions have been included in `R/regular_qk_grid.R` that wrap the three major functions of the workflows described in `quadkey_identified_data_to_raster` (`add_regular_polygon_grid()`) and `Facebook_mobility_csvs_to_raster_files.`(`get_regular_polygon_grid()`) . Fernandez-Lab-WSU/quadkeyr@3e354aa6
- Testing coverage has been increased from 84% to 95%.

### **IMPROVEMENTS IN PREEXISTENT FUNCTIONS**
- Error messages are now included in functions that take QuadKeys as arguments. Now, it is checked that the QuadKeys are strings and composed only of the digits '0', '1', '2', or '3'. All these messages have corresponding tests. Fernandez-Lab-WSU/quadkeyr@1f4e429
- Tests for `polygon_to_raster()` are now available. Fernandez-Lab-WSU/quadkeyr@9eac5fbd
- In order to create `add_regular_polygon_grid()`, the function `regular_qk_grid()` can now add rows to a data.frame when adding the extra grid while keeping other variables of interest.
- The parameter "Level of detail" has been replaced by "zoom level" in all package functions. Fernandez-Lab-WSU/quadkeyr@1f4e429
- Some functions have changed names to more descriptive ones, for example `create-raster()` to `create_stars_raster()` Fernandez-Lab-WSU/quadkeyr@f105177c
- The code style has been improved; unnecessary spaces have been removed, vignettes are now linked in Roxygen2, and examples have been updated. Additionally, the text has been enhanced for clarity. Fernandez-Lab-WSU/quadkeyr@f105177c
- `get_qk_tile()` is now wrapped inside `get_qk_coord()`, facilitating its use for other internal functions such as `regular_qk_grid()`. Fernandez-Lab-WSU/quadkeyr@605476993
- It is no longer necessary to add +1 to use the correct number of columns and rows in create_stars_raster(), as there have been changes in `create_qk_grid()`. Fernandez-Lab-WSU/quadkeyr@a7a5f32e
- The variable name changed from `time` to `hour` has been completed in all the related functions and datasets. Fernandez-Lab-WSU/quadkeyr@4bd5cc95

### **APP**
* Thanks to the new function `quadkey_to_polygon()` now it is possible to visualize the QuadKey as a point and a polygon in the app. [f0ba3ad](https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/1f4e429fff0a9ec7c04f5bc08d068ff084e1e1b9)

## REVIEW @mpaulacaldas
**Review Comments**
Thank you very much for the opportunity to review this package. As someone who has struggled with QuadKeys in the past, I think this package addresses a very real gap in the R spatial software space, and I believe it will be useful to many social scientists.
I leave below detailed comments on the different aspects of the package. The review is quite extensive, so please don’t hesitate to ignore the more minor suggestions. Likewise, I’m happy to discuss any points that are unclear or help brainstorm solutions.

### 1 - **README**
This is an excellent README. If I could provide only one minor comment, it would be to put the section “What can this package do for you?” before “What are quadkeys and tile maps?”

> This one is easy to see in the README: https://github.com/Fernandez-Lab-WSU/quadkeyr
Fernandez-Lab-WSU/quadkeyr@6e003884

### 2 - **Vignettes**
Overall, I find the vignettes are informative and easy to understand, with truly excellent pedagogical figures.
I leave below my comments for possible improvements, to complement some of the points raised previously by Vincent:

**On C-from_facebook_quadkey_data_to_raster.Rmd:**

- The build of this vignette is failing as of[ Fernandez-Lab-WSU/quadkeyr@183788c](https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/183788c0eb24d47b5da06fe176680db27e4da886) (the commit I am using for this review). This is because the data frame passed to apply_weekly_lag() does not contain an hour column, that it needs for it to run.

> This issue has been resolved; some sections were not updated despite not giving errors. I missed the command that allows me to change all the variables in a project, which I have used with other editors. Next time, I will use regex. I am trying to make small commits for easier review, so this might not have helped either, I was expecting you to use the same commit as Vincent.
This is the function [now](https://github.com/Fernandez-Lab-WSU/quadkeyr/tree/9b7b8c438933022ed81fc1f102e162bff556917a) and here is one of the commits I used to fix it Fernandez-Lab-WSU/quadkeyr@4bd5cc9.

- As I understand it, the objective of this vignette is to showcase how to create a raster from a table where cells are identified with quadkeys. To do so, you use Facebook mobility data as an example, but the logic can be applied to other quadkey-identified datasets. With this in mind, I think it would be useful to have a more general title, for example: “Converting a quadkey-identified data set to a raster”. In the first sentences of the vignette, you can mention that your example will use FB mobility data, but the same logic could be applied to any data set with column with quadkey-identifiers.

> That is the new name of the vignette now, as you can see in the vignette: https://fernandez-lab-wsu.github.io/quadkeyr/articles/quadkey_identified_data_to_raster.html
This is the vignette [in the present commit](https://github.com/Fernandez-Lab-WSU/quadkeyr/blob/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes/quadkey_identified_data_to_raster.Rmd)

- The example in the vignette is not reproducible, as it depends on the files in "../geocovid/data/rasters/city/" – a relative path only available in the developer’s computer. I see two ways to fix this issue:
If the data is small enough, add it to the package under inst and access it using system.file() or fs::path_package(), as suggested in the[ data chapter](https://r-pkgs.org/data.html#sec-data-system-file) of the R packages book.
If the data is too large, create a helper that downloads the files into a directory and reads them from there (FB mobility data may be found under a stable link).

> Thank you for your detailed suggestions, I agree. All the examples of the vignettes are reproducible now, as there have been added example files in `inst/extdata`.
[c27e452](https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/c27e452800f38c3d0323c9f4a2bc108e16a33a7e)

- In steps 2 and 3, it would be useful to explicitly mention in the text that the number of quadkeys went from 150 to 396 (though the info is available in the printed sf object, it is a bit difficult to parse). Explicitly mentioning these numbers may make it easier for readers to understand what you mean by “completing the grid”.
> Done in [6e00388](https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/6e003884e46e1fdbd0ccd42254a53b10cdfc649a)

- Seeing step 4, it is unclear to me why we bother expanding the grid, if in the end the inner_join() will get rid of all the quadkeys from polygrid that are implicitly missing from files. Perhaps it would be useful to re-instate that the grid expansion is performed because creating the polygons (and later the rasters) in this implementation requires a full grid.
> There are 3 reasons for this approach:
> 1. **Flexibility with Facebook Mobility Datasets**:  When receiving a set of Facebook mobility datasets, it's uncertain whether exactly the same QuadKeys will be reported in all of them. To address this, I detect all the QuadKeys reported after reading all the Facebook data frames and convert them to polygons, creating a data.frame containing only the columns `quadkey` and `geometry` that you can later join with the Facebook mobility data. This approach not only streamlines the process for the current datasets but also prepares to process potential future files with unknown QuadKeys for the studied location.
> Alternatively, users can create the QuadKey polygon grid using the bounding box of the area of interest and then join the data. This option is facilitated by the function create_qk_grid(), and both approaches are now documented in the vignette:
> 2. **Computational cost** - Also, consider the computational cost associated with calculating polygons for each file, especially when many of the files share almost identical QuadKeys. By aggregating and storing the QuadKeys separately, we avoid redundant calculations for overlapping QuadKeys across multiple Facebook files.
> 3. **Grid Creation for Raster**: The third reason is the necessity of the grid for raster operations, as you mentioned.
I've also introduced the function `quadkey_df_to_polygon()`, which allows conversion without completing the grid. I explicitly mention this difference to the user to ensure clarity on the approach being used.
> You can read the improvements in the new versions of the vignettes that correspond to [this commit](https://github.com/Fernandez-Lab-WSU/quadkeyr/tree/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes)

- The resolution of the last image is off in my display using the RStudio viewer.
> I am not very sure about which one is the last image, but I have improved how all the images are rendered, so I hope it is solved.

**On B-from_facebook_quadkey_data_to_raster.Rmd:**
     - I think the name/title of this vignette could still be improved. I would suggest “Building a quadkey grid from a bounding box” (e.g. and a file name like quadkey_grid_from_bbox.Rmd for example) or something along those lines. In the end, the main difference between this vignette and C-from_facebook_quadkey_data_to_raster is that you are showing how to build your own quadkey grid over a given extent, but all other steps (i.e. going from a data set with quadkeys to a raster) are the same. I imagine most users will end up having a workflow like C-from_facebook_quadkey_data_to_raster, but only few will need to build their own quadkey grid.
I assume xmin, xmax, ymin and ymax refer to the coordinates of the bounding box, following the common notation from sf. If so, I would advise using the term bounding box explicitly, since it is a term that most spatial R users will be familiar with.

> This vignette suffers the biggest changes. It is what I am saying in the previous point, you can use the bounding box or the list of QuadKeys to calculate this.
> You can read the improvements in the new versions of the vignettes that correspond to [this commit](https://github.com/Fernandez-Lab-WSU/quadkeyr/tree/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes)

- In this vignette and throughout the package in general, it would be better to use the word zoom level explicitly instead of “level of detail”. Zoom level is the term that package users are most likely to look for when they are skimming the documentation, and it is also the parameter name used in other packages (e.g. slippymath)

> I tried to keep it exactly as appears in the documentation of Microsoft Bing Tile Maps, but I agree that the connection with other packages is desirable. Done in [this commit](https://github.com/Fernandez-Lab-WSU/quadkeyr/commit/1f4e429fff0a9ec7c04f5bc08d068ff084e1e1b9) and I also added a note in the first vignette about both terms being identical.

**On A-bing_map_tile_system_functions.Rmd:**
- This is a minor note, but I think it’s easier (and still correct) to say “quadkey” instead of “quadkey number”.

- Bravo for including the caveat in Section 3. Perhaps a more eye-catching title would help users understand the importance of reading this section (e.g. “Caveats when converting coordinates”)
> Done, and also I included a call in Roxygen to the vignette.
> You can read the improvements in the new versions of that vignette that correspond to [this commit](https://github.com/Fernandez-Lab-WSU/quadkeyr/blob/9b7b8c438933022ed81fc1f102e162bff556917a/vignettes/quadkey_to_sf_conversion.Rmd)

**On D-quadkey_visualization_app.Rmd:**
It’s a nice and valuable addition to have a vignette introducing the Shiny app, with screenshots showing common behaviors. I will leave here some minor improvements I think you could make to the app:
- I think it would be useful if the Quadkey tab, showed the boundaries of the user-specified quadkey in addition to only the upper-left corner. As an analyst, I often want to visualize the area covered by a specific quadkey more than the coordinates of its extreme points. If you introduce these changes, I think a better name for the tab could be “Visualising a quadkey”.
- I think the Grid tab could have a better name, so users know that its goal is to produce a quadkey grid from a set of coordinates and a zoom level. I would suggest “Creating a quadkey grid”.
> This how the app

- Lastly, and going slightly against[ @vincentvanhees](https://github.com/vincentvanhees) suggestion (sorry!) I think you should remove the alphabetical tags from the vignettes names. This is because locally, users access vignettes using vignette("file_name_without_extension"), and longer file names can be harder to type and remember. As someone having working with quadkeys in the past, the most useful vignette seems to be C-from_facebook_quadkey_data_to_raster, so if you can, I would prioritise making that specific name shorter, while still informative (not an easy task at all!)
> Done

### 2 - **Function documentation**
Following on the above, it would be useful to add to the function documentation of apply_weekly_lag(), a line in the Description clarifying that these are specific to the format of FB mobility data.

Is quadkeyr::clip() supposed to be exported? Its documentation mentions that this is an internal function and doesn’t provide much detail on functionality. It also masks graphics::clip(), so if you don’t think it will be much use to end-users, I would suggest to keep it internal.If you want to roxygen to produce a documentation file for the function, while preventing the function from being exported, you can use the @keywords internal tag instead of @export.
> Yes, I thought about this. Most of the functions in `map_functions.R` could be internal. As I wasn’t sure how potential users could use the package I decided to make available all the functions. However, it is true that among the functions in `map_functions.R`, `clip()` is the most difficult to imagine and use in isolation and the masking is giving a justification to make this one internal.
> Done in  Fernandez-Lab-WSU/quadkeyr@c8f97282d

Similar comment for `complete_grid_for_polygons()`. Is this function meant to be exported? Are you exporting it because it may be useful for advanced users? If so, it would be useful to mention it in the function documentation.

> I agree, this is pending.

The code style is at times inconsistent across the documentation (examples and vignettes). For example, = and <- both being used for assignment, sometimes even within the same example (e.g. create_raster()). Using styler::style_pkg() to catch these may be helpful.
> Done in  Fernandez-Lab-WSU/quadkeyr@f105177c. This also fixes most of the space issues that @vincentvanhees mentioned.

What is the difference between extract_qk_coord() and extract_tile_coord()? It is unclear from the documentation and the examples. Doing a quick test, I also get that both functions return the same result, with the exception that extract_qk_coord() does not require a level argument.
```
 grid <- create_qk_grid(
   xmin = -59,
   xmax = -40,
   ymin = -38,
   ymax = -20,
   level = 6
 )

 tile_result <- extract_tile_coord(data = grid$data, level = 6)
 qk_result   <- extract_qk_coord(data = grid$data)

 waldo::compare(tile_result, qk_result)
 #> ✔ No differences
```
> Yes, the difference between these two functions is the input, not the output. Instead of using the QuadKey to get the geographic coordinates, you can use the Tile coordinates.
One of the QuadKeys’ properties is that the length of the string provides information about the level of detail, that is why you need to specify the level when starting the conversion from the Tile coordinates.
What I think you find weird about this is that I don’t need to define in detail both functions, I can call `get_tile_coord()` inside `get_qk_coord()`. That is the change I applied in Fernandez-Lab-WSU/quadkeyr@6054769

The example in format_fb_data() is not reproducible for users and is wrapped in a dontrun{}. Couldn’t you provide an internal data frame mimicking the format of facebook mobility data? The name fb_mobility_file that you use in the function documentation sounds like a great name for this example data frame. Alternatively, you could also demo how to download the file from a stable link with download.file(), saving to a temporary directory, and reading from there.
> Done in Fernandez-Lab-WSU/quadkeyr@63c8932

In ground_res(), it would be useful to add a link to the Microsoft documentation. Same goes for mapscale(), mapsize().
> Done in  Fernandez-Lab-WSU/quadkeyr@2772dd9

In latlong_to_quadkey(), do the returned coordinates correspond to the upper-left corner of the pixel? Is this where the caveat from the vignette applies? If so, it wouldn’t hurt to repeat it again in the function documentation. Same comment goes for pixelXY_to_latlong(), tileXY_to_pixelXY()
> Done in Fernandez-Lab-WSU/quadkeyr@157701dba0

[…]
