### 2024-02-14 #619 @flor14
@emilyriederer and reviewers

There are 4 comments and 1 bug on the last revision that I have pending, but there are minor ones:
1 - Make the function `complete_grid_to_polygon()` internal
2 - Split the reference section regarding what the functions are for.
3 - I want to see if the argument names of quadkey_* functions still make sense after all the changes.
4- Fix a bug in one of the vignettes, but the `hour` column appears as NA by a parsing error in `read.csv`.

I have to leave this for some days now, so I decided to share my comments anyway. I will open issues with these pending items and post them when finished them.
