### 2024-02-20 0fa24de
Update docs and fix bug in read_fb_mobility_files() (ropensci/software-review#619) - Fix bug in read_fb_mobility_files() that was producing parsing issues and update docs - Add contributing.md link to README and other minor updates - Fix url breaks in the roxygen docs
