+++ .github/CONTRIBUTING.md
@@ -45,3 +45,4 @@ If you’ve found a bug, please file an issue that illustrates the bug with a mi
 Please note that the quadkeyr project is released with a
 [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this
 project you agree to abide by its terms.
+
