### 2024-02-20 7d0a9da
Add changes in quadkey_to_latlong() and update docs: - Modify quadkey_to_latlong() argument to quadkey_data - Include new diagrams in vignettes for explain quadkey_to_latlong() and latlong_to_quadkey() - Fix urls in roxygen (again) - Add other fixes in the vignette's yaml (toc-depth) and missing examples in roxygen docs
