### 2024-02-20 0df053e
Run styler::style_pkg() ropensci/software-review#619
