+++ R/grid_to_polygon.R
@@ -35,7 +35,7 @@ grid_to_polygon <- function(data) {
   if (!("sf" %in% class(data))) {
     stop("The dataset should be of class 'sf'")
   }
-  
+
   # Convert the QuadKeys to tile coordinates
   # if that columns aren't present in the data
   # if (!("tileX" %in% colnames(data) | "tileY" %in% colnames(data))) {
@@ -43,77 +43,80 @@ grid_to_polygon <- function(data) {
   #     "The 'tileX' and 'tileY' columns have been generated",
   #     "using the 'quadkey_to_tileXY' function."
   #   ))
-  
+
   # In the case that one of this columns is not present,
   # calculate both again.
   data$tileX <- NA
   data$tileY <- NA
-  
+
   data <- data |>
     dplyr::rowwise() |>
-    dplyr::mutate(tileX = quadkey_to_tileXY(.data$quadkey)$tileX,
-                  tileY = quadkey_to_tileXY(.data$quadkey)$tileY)
-  
-  
+    dplyr::mutate(
+      tileX = quadkey_to_tileXY(.data$quadkey)$tileX,
+      tileY = quadkey_to_tileXY(.data$quadkey)$tileY
+    )
+
+
   # We add the extra row and column we need for the grid and
   # Convert it to an sf POINT data.frame
   extragrid <- complete_grid_for_polygons(data)
   extragrid <- get_tile_coord(extragrid,
-                              zoom = unique(nchar(data$quadkey)))
-  
+    zoom = unique(nchar(data$quadkey))
+  )
+
   # Now we can combine it with the original data.
   if (ncol(data) == ncol(extragrid)) {
     # combines the new data with the extended grid of points
     data <- rbind(data, extragrid)
-  } else{
+  } else {
     # In the case that there are more columns in data,
     # keep the rows but adding NAs in the extragrid cells.
     data <- dplyr::bind_rows(data, extragrid)
   }
-  
+
   polydata <- c() # https://github.com/r-spatial/sf/issues/354
-  
+
   # The QuadKeys of interest are the ones that are not NA:
   # the original QuadKeys of the grid
-  subdata <- subset(data,!is.na(data$quadkey))
-  
+  subdata <- subset(data, !is.na(data$quadkey))
+
   for (i in seq_len(nrow(subdata))) {
-    x <- subdata[i,]$tileX
-    y <- subdata[i,]$tileY
-    
+    x <- subdata[i, ]$tileX
+    y <- subdata[i, ]$tileY
+
     # This point will always be a QuadKey in the dataframe
-    a <- data[data$tileX == x & data$tileY == y,]
-    
+    a <- data[data$tileX == x & data$tileY == y, ]
+
     # b, c and d can be part of the extended grid
-    b <- data[data$tileX == x & data$tileY == (y + 1),]
-    
-    c <- data[data$tileX == x + 1 & data$tileY == y,]
-    
-    d <- data[data$tileX == x + 1 & data$tileY == y + 1,]
-    
+    b <- data[data$tileX == x & data$tileY == (y + 1), ]
+
+    c <- data[data$tileX == x + 1 & data$tileY == y, ]
+
+    d <- data[data$tileX == x + 1 & data$tileY == y + 1, ]
+
     polygon <- rbind(a, b, c, d) |>
       sf::st_bbox() |>
       sf::st_as_sfc()
-    
-    subdata_row <- subdata[i,] |>
-      dplyr::mutate(geometry = sf::st_sfc(polygon))  |>
+
+    subdata_row <- subdata[i, ] |>
+      dplyr::mutate(geometry = sf::st_sfc(polygon)) |>
       sf::st_set_geometry("geometry")
-    
+
[…]
