### 2024-02-20 c9fffb7
Include reference section in _pkgdown.yaml (ropensci/software-review#619) - Organize functions in relation to the different vignettes / uses - Fix title roxygen documentation - Fix bug vignette due to changes in `quadkey_to_latlong()`
