+++ man/mapsize.Rd
@@ -22,6 +22,7 @@ System documentation.
 \examples{
 
 mapsize(zoom = 6)
+
 }
 \references{
 \url{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}
