### 2024-02-20 #619 @flor14
Vincent, Maria Paula, and @emilyriederer,

First of all, thank you for your patience. Here are the last modifications, the revision is now complete on my end:

- I have added a `references` section to the `_pkgdown.yaml`, using the same names as the vignettes for the titles. All the functions that are not listed here because they weren't explicitly mentioned in the vignettes, have been classified as internal including the datasets and `get_tile_coord()`.
This commit addresses @mpaulacaldas comment about the references section and also clarifies what functions are available for their use.
👉🏼 Fernandez-Lab-WSU/quadkeyr@c9fffb789

- I kept thinking about the `quadkey_*` functions' arguments and found it challenging to get an illustrative argument name for a function that can accept strings or vectors, in comparison to the ones that only accept strings.
I have implemented two modifications regarding this:

1 - I have created an image illustrating this concept and added it to the vignette. Aditionally, I have created it for `latlong_to_quadkey()` as well:

2 - I have updated the argument names for `quadkey_to_latlong()` to `quadkey_data` and keep `quadkey` for `quadkey_to_tileXY()`.
👉🏼Fernandez-Lab-WSU/quadkeyr@7d0a9da17

- The bug I mentioned in the previous comment was fixed here 👉🏼 Fernandez-Lab-WSU/quadkeyr@0fa24de46

Finally, I have run again `stylr_pkg()` after most of this changes 👉🏼 Fernandez-Lab-WSU/quadkeyr@0df053ebb

---

Last but not least, I would like to thank you for the detailed review once again. When I began this particular part of the project in August 2023, I was very new to working with QuadKeys. It has been a very ambitious package from the start, but as the intention is to continue working with this data I know that all of this hard work will pay off.
I am quite happy with the results, as we now have a package to condense all that I've learned from my research and from you to share with the rest of the community.

Question: Would it be possible to leave a version of this package in Spanish (I mean, not to have the review in Spanish again, if not leave a version of it in Spanish)? I am not sure if that is possible or if you have any comments/examples about other developers doing something like that.

Thank you! And just in case @ropensci-review-bot check package
