### 2024-02-21 #619 @emilyriederer
Thank you @flor14 ! Your dedication to this ambitious project is very impressive and its great to see all these updates. I know we are actively reviewing packages now in multiple languages, so I will check in with the editorial team regarding best practices for maintaing versions of both.

@mpaulacaldas and @vincentvanhees - can you please take a pass through all these wonderful updates and confirm your approval with [this template](https://devguide.ropensci.org/approval2template.html)?
