### 2024-02-22 e8694e8
Fix small typo in DESCRIPTION
