+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   quadkey_identified_data_to_raster: quadkey_identified_data_to_raster.html
   quadkey_to_sf_conversion: quadkey_to_sf_conversion.html
   quadkey_visualization_app: quadkey_visualization_app.html
-last_built: 2024-02-20T21:12Z
+last_built: 2024-02-22T16:16Z
 
