### 2024-02-28 df63bf8
Add argument to format_fb_data
