### 2024-03-01 #619 @emilyriederer
Hi there @mpaulacaldas and @vincentvanhees - just a reminder, at your convenience, could you please review @flor14 's updates and confirm your approval with [this template](https://devguide.ropensci.org/approval2template.html)? Per our [reviewer guide](https://devguide.ropensci.org/reviewerguide.html#followupreviewer), the goal of this stage is to ensure you feel your comments have been sufficiently addressed
