### 2024-03-02 e3af68a
Update apply_weekly_lag() and add tests - Add missing print() in polygon_to_raster.R function - Minor chnages in the docs
