### 2024-03-02 #619 @flor14
For the past few days, I have been using `quadkeyr` and I have added improvements in `apply_weekly_lag()`,  `read_fb_mobility_files()`, and `format_fb_data()` along with their tests.
- The loop in `apply_weekly_lag()` was removed. Reducing the amount of loops was one of Vincent's comments.
- `read_fb_mobility_files()` and `format_fb_data()` now have an additional argument, making them more flexible for working with real data.
All these changes are available in the `read-and-7-day-lag` branch: https://github.com/Fernandez-Lab-WSU/quadkeyr/commits/read-and-7-day-lag/

I will not merge these changes because I believe the review has already been addressed, and I don't want to create confusion. However, if after reading the review you still have concerns about these specific functions, you can check that branch.

Thank you.
