### 2024-03-03 #619 @vincentvanhees
## Reviewer Response

Great job @flor14! My comments have been addressed satisfactory.

A few remarks for your consideration:
- Going back to our [correspondence](https://github.com/ropensci/software-review/issues/619#issuecomment-1904495672) about the `html_document` versus `html_vignette`: I learnt last week that what I suggested (html_document) is less lightweight than (html_vignette) according to [bookdown documentation](https://bookdown.org/yihui/rmarkdown/r-package-vignette.html). So, if this is a concern when the package grows in the future then you may want to reverse to `html_vignette`. For now I guess it is not a problem.

- The DESCRIPTION file lists you as author and creator but does not list copyright holders (`cph`). If this is a hobby project then it is also you, but if someone paid for your time then usually they are the copyright holder. If applicable, listing them gives credit for their support and is helpful to know for potential contributors.

- I am not familiar with rOpenSci's release process, but if the acceptance of this package does not generate a tagged release I would recommend creating one in the near future, such that you have a clear reference point for this stage of the development. Don't forget to also describe this as your first release in NEWS.md.

#### Final approval (post-review)

Estimated hours spent reviewing: 1
