### 2024-03-03 #619 @mpaulacaldas
## Reviewer Response

Thank you @flor14 for all of the detailed and careful changes you made in this last round. With these, I believe you have addressed my key comments in a satisfactory way, and I am happy for to recommend approval.

Like Vincent, I leave here a couple of remarks for future consideration:

- I think you may have some confusion on the use of `@export` and the `@keywords internal` tag. In general, you do not use `@keywords internal` for functions that you export with `@export`, but I see for example that you are doing so for `get_tile_coord()` and `complete_grid_for_polygons()`. If you don't intend to export these functions (in other words, you want them to be internal), then you should remove the `@export` tag. If on the contrary, you want to export the functions, then you don't need the `@keywords internal` tag, since Roxygen builds by default documentation for all exported functions. Right now, you are in strange situation where those two functions _are_ exported, but their documentation is _hidden_ in the pkgdown website.

- A minor bug, but the images in your README are not showing up in the pkgdown website.

#### Final approval (post-review)

Estimated hours spent reviewing: 1
