### 2024-03-04 3106f36
Fix figure pkgdown index and remove @export from internal functions (ropensci/software-review#619)
