+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
@@ -6,5 +6,5 @@ articles:
   quadkey_identified_data_to_raster: quadkey_identified_data_to_raster.html
   quadkey_to_sf_conversion: quadkey_to_sf_conversion.html
   quadkey_visualization_app: quadkey_visualization_app.html
-last_built: 2024-02-23T21:30Z
+last_built: 2024-03-05T16:10Z
 
