### 2024-03-06 #619 @flor14
I am on hold with this because the researcher I am working with is asking the university about the copyright. Also, I will ask a question in ROpenSci Slack regarding how to specify the authorship, in case you have an opinion about it.
