### 2024-03-07 #619 @vincentvanhees
My understanding is that if you work for an institute then usually they are the copyright holder and not the external funder that funded the research.

> Also, I will ask a question in ROpenSci Slack regarding how to specify the authorship, in case you have an opinion about it.

You can just do:

```
Authors@R:
    c(person(given = "Florencia",
           family = "D'Andrea",
           email = "florencia.dandrea@gmail.com",
           role = c("aut", "cre"),
           comment = c(ORCID = "0000-0002-0041-097X")),
    person("Your Institute Name", role="cph"))
```
