### 2024-03-07 #619 @flor14
Thank you @vincentvanhees, you are always ready to answer.
The researcher told me that she wanted to ask someone from the university about this, so I will wait anyway. I don't work for any institution, this is just a contract.
