### 2024-03-13 8e33774
Update DESCRIPTION authors and fix bug in pixelXY_to_latlong()
