### 2024-03-13 b3bd37c
Fix package version in authors (DESCRIPTION)
