### 2024-03-13 9fe62bd
Update pkgdown site, DESCRIPTION, and CITATION - Fix typos - Update NEWS - Minor text improvements in vignettes
