+++ README.Rmd
@@ -8,12 +8,11 @@ output: github_document
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
+  fig.path = "man/figures/README-"
 )
 ```
 
-# `quadkeyr` <img src="vignettes/logo.png" align="right" height="200" />
+# quadkeyr <img src="vignettes/logo.png" align="right" height="100"/>
 
 <!-- badges: start -->
 [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental) [![test-coverage](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/test-coverage.yaml/badge.svg)](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/test-coverage.yaml) [![R-CMD-check](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/check-release.yaml/badge.svg)](https://github.com/Fernandez-Lab-WSU/quadkeyr/actions/workflows/check-release.yaml)
@@ -21,9 +20,6 @@ knitr::opts_chunk$set(
 [![Codecov](https://img.shields.io/codecov/c/github/Fernandez-Lab-WSU/quadkeyr)](https://codecov.io/gh/Fernandez-Lab-WSU/quadkeyr)
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/619_status.svg)](https://github.com/ropensci/software-review/issues/619)
 <!-- badges: end -->
-
-⚠️ This package is under review
-
 
 ### What can this package do for you?
 The `quadkeyr` R package presents a comprehensive toolkit 
@@ -33,7 +29,6 @@ Designed to integrate Quadkey-Identified data into R workflows,
 this package facilitates the creation of QuadKey grids and raster images
 and introduces specialized functions for
 the processing of [Facebook Mobility data](https://dataforgood.facebook.com/).
-
 
 ### What are QuadKeys in Tile Maps?
 Tile maps divide the Earth's surface into a grid of tiles, 
