### 2024-03-13 04c15d2
Remove docs - internal functions - Add figure to articles/ to fix bug
