+++ R/create_stars_raster.R
@@ -15,7 +15,6 @@
 #'
 #' @return A `stars` object.
 #' @export
-#'
 #' @examples
 #'
 #' # Basic workflow
