+++ man/clip.Rd
@@ -1,35 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/map_functions.R
-\name{clip}
-\alias{clip}
-\title{Clips a number to the specified minimum and maximum values.}
-\usage{
-clip(n, min_value, max_value)
-}
-\arguments{
-\item{n}{The number to clip.}
-
-\item{min_value}{Minimum allowable value.}
-
-\item{max_value}{Maximum allowable value.}
-}
-\value{
-The clipped value.
-}
-\description{
-This function is user internally by \code{\link[=latlong_to_pixelXY]{latlong_to_pixelXY()}} and
-\code{\link[=pixelXY_to_latlong]{pixelXY_to_latlong()}}.
-For further information, refer to the Microsoft Bing Maps Tile System
-documentation.
-}
-\examples{
-\dontrun{
-quadkeyr::clip(
-  n = 1,
-  min_value = 3,
-  max_value = 5
-)
-}
-
-}
-\keyword{internal}
