### 2024-03-13 #619 @flor14
The review is finished, I have updated the authors, the`NEWS.md` and created a [release](https://github.com/Fernandez-Lab-WSU/quadkeyr/releases).

Remaining questions:
- What about the submission to JOSS? Should I mention this issue with the submission?
- The idea is to leave a version in Spanish available as well, any comment on how to do that is welcomed.

Thank you
