+++ README.Rmd
@@ -45,7 +45,7 @@ mapping applications.
 
 
 ```{r qks, echo = FALSE, out.width= "70%", fig.align='center'}
-knitr::include_graphics("vignettes/bing_quadkeys.jpg")
+knitr::include_graphics("https://github.com/Fernandez-Lab-WSU/quadkeyr/blob/main/docs/articles/bing_quadkeys.jpg?raw=true")
 ```
 <p style="text-align: center">The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft's Bing Maps Tile System webpage.</p>
 
