+++ DESCRIPTION
@@ -17,9 +17,9 @@ Authors@R:
            role = c("rev"),
            comment = c("Maria Paula Caldas reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619")),
     person(given = "Vincent",
-           family = "Van Hess",
+           family = "van Hees",
            role = c("rev"),
-           comment = c("Vincent Van Hess reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619")),
+           comment = c("Vincent van Hees reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619")),
     person("CDC's Center for Forecasting and Outbreak Analytics",
             comment = c("This project was made possible by cooperative agreement CDC-RFA-FT-23-0069 (grant # NU38FT000009-01-00) from the CDC's Center for Forecasting and Outbreak Analytics. Its contents are solely the responsibility of the authors and do not necessarily represent the official views of the Centers for Disease Control and Prevention."),  role = c("fnd")),
     person("MIDAS-NIH COVID-19 urgent grant program", role = c("fnd")),
