### 2024-03-14 #619 @emilyriederer
A few additional notes @flor14 :

- I'd strongly encourage you to market your package with a post as described in one of the last paragraphs of the automated post above. I think this package will have a lot of interest and I'd love to get the word out there
- For JOSS, you can use the rOpenSci repository link once you've transferred it and potentially link to this review in your submission. As you can see in [this recent submission](https://github.com/openjournals/joss-reviews/issues/5885), their system is set up to notice
- On multilanguage docs, this is still an open issue. [This package](https://github.com/mpru/karel) is currently under review and strives to have multilanguage docs, so you can check out their set up. There's also the [babelquarto](https://docs.ropensci.org/babelquarto/) package for multilingual websites but it doesn't directly work with `pkgdown`. Meanwhile [this project](https://github.com/eliocamp/rhelpi18n) is in development to do something similar for help pages, but to my understanding it's not a stable release yet
