### 2024-03-15 #619 @yabellini
I will jump here to add that we are discussing multilingual documentation and publishing and sharing projects, ideas, and progress in the #multilingual channel in our Slack. 😃
